Split docker_cli_build_test.go into several pieces
