+++ integration-cli/docker_cli_build_test.go
@@ -1,9 +1,7 @@
 package main
 
 import (
-	"archive/tar"
 	"bufio"
-	"bytes"
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
@@ -16,11 +14,9 @@ import (
 	"runtime"
 	"strconv"
 	"strings"
-	"text/template"
 	"time"
 
 	"github.com/docker/docker/builder/dockerfile/command"
-	"github.com/docker/docker/pkg/archive"
 	"github.com/docker/docker/pkg/integration/checker"
 	"github.com/docker/docker/pkg/stringutils"
 	"github.com/go-check/check"
@@ -85,217 +81,6 @@ func (s *DockerSuite) TestBuildShCmdJSONEntrypoint(c *check.C) {
 
 	if strings.TrimSpace(out) != "/bin/sh -c echo test" {
 		c.Fatalf("CMD did not contain /bin/sh -c : %s", out)
-	}
-
-}
-
-func (s *DockerSuite) TestBuildEnvironmentReplacementUser(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildenvironmentreplacement"
-
-	_, err := buildImage(name, `
-  FROM scratch
-  ENV user foo
-  USER ${user}
-  `, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	res, err := inspectFieldJSON(name, "Config.User")
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	if res != `"foo"` {
-		c.Fatal("User foo from environment not in Config.User on image")
-	}
-
-}
-
-func (s *DockerSuite) TestBuildEnvironmentReplacementVolume(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildenvironmentreplacement"
-
-	_, err := buildImage(name, `
-  FROM scratch
-  ENV volume /quux
-  VOLUME ${volume}
-  `, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	res, err := inspectFieldJSON(name, "Config.Volumes")
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	var volumes map[string]interface{}
-
-	if err := json.Unmarshal([]byte(res), &volumes); err != nil {
-		c.Fatal(err)
-	}
-
-	if _, ok := volumes["/quux"]; !ok {
-		c.Fatal("Volume /quux from environment not in Config.Volumes on image")
-	}
-
-}
-
-func (s *DockerSuite) TestBuildEnvironmentReplacementExpose(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildenvironmentreplacement"
-
-	_, err := buildImage(name, `
-  FROM scratch
-  ENV port 80
-  EXPOSE ${port}
-  ENV ports "  99   100 "
-  EXPOSE ${ports}
-  `, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	res, err := inspectFieldJSON(name, "Config.ExposedPorts")
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	var exposedPorts map[string]interface{}
-
-	if err := json.Unmarshal([]byte(res), &exposedPorts); err != nil {
-		c.Fatal(err)
-	}
-
-	exp := []int{80, 99, 100}
-
-	for _, p := range exp {
-		tmp := fmt.Sprintf("%d/tcp", p)
-		if _, ok := exposedPorts[tmp]; !ok {
-			c.Fatalf("Exposed port %d from environment not in Config.ExposedPorts on image", p)
-		}
-	}
-
-}
-
-func (s *DockerSuite) TestBuildEnvironmentReplacementWorkdir(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildenvironmentreplacement"
-
-	_, err := buildImage(name, `
-  FROM busybox
-  ENV MYWORKDIR /work
-  RUN mkdir ${MYWORKDIR}
-  WORKDIR ${MYWORKDIR}
-  `, true)
-
-	if err != nil {
-		c.Fatal(err)
-	}
-
-}
-
-func (s *DockerSuite) TestBuildEnvironmentReplacementAddCopy(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildenvironmentreplacement"
-
-	ctx, err := fakeContext(`
-  FROM scratch
-  ENV baz foo
-  ENV quux bar
-  ENV dot .
-  ENV fee fff
-  ENV gee ggg
-
-  ADD ${baz} ${dot}
-  COPY ${quux} ${dot}
-  ADD ${zzz:-${fee}} ${dot}
-  COPY ${zzz:-${gee}} ${dot}
-  `,
-		map[string]string{
-			"foo": "test1",
-			"bar": "test2",
-			"fff": "test3",
-			"ggg": "test4",
-		})
-
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-
-}
-
-func (s *DockerSuite) TestBuildEnvironmentReplacementEnv(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildenvironmentreplacement"
-
-	_, err := buildImage(name,
-		`
-  FROM busybox
-  ENV foo zzz
-  ENV bar ${foo}
-  ENV abc1='$foo'
-  ENV env1=$foo env2=${foo} env3="$foo" env4="${foo}"
-  RUN [ "$abc1" = '$foo' ] && (echo "$abc1" | grep -q foo)
-  ENV abc2="\$foo"
-  RUN [ "$abc2" = '$foo' ] && (echo "$abc2" | grep -q foo)
-  ENV abc3 '$foo'
-  RUN [ "$abc3" = '$foo' ] && (echo "$abc3" | grep -q foo)
-  ENV abc4 "\$foo"
-  RUN [ "$abc4" = '$foo' ] && (echo "$abc4" | grep -q foo)
-  `, true)
-
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	res, err := inspectFieldJSON(name, "Config.Env")
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	envResult := []string{}
-
-	if err = unmarshalJSON([]byte(res), &envResult); err != nil {
-		c.Fatal(err)
-	}
-
-	found := false
-	envCount := 0
-
-	for _, env := range envResult {
-		parts := strings.SplitN(env, "=", 2)
-		if parts[0] == "bar" {
-			found = true
-			if parts[1] != "zzz" {
-				c.Fatalf("Could not find replaced var for env `bar`: got %q instead of `zzz`", parts[1])
-			}
-		} else if strings.HasPrefix(parts[0], "env") {
-			envCount++
-			if parts[1] != "zzz" {
-				c.Fatalf("%s should be 'foo' but instead its %q", parts[0], parts[1])
-			}
-		} else if strings.HasPrefix(parts[0], "env") {
-			envCount++
-			if parts[1] != "foo" {
-				c.Fatalf("%s should be 'foo' but instead its %q", parts[0], parts[1])
-			}
-		}
-	}
-
-	if !found {
-		c.Fatal("Never found the `bar` env variable")
-	}
-
-	if envCount != 4 {
-		c.Fatalf("Didn't find all env vars - only saw %d\n%s", envCount, envResult)
 	}
 
 }
@@ -388,248 +173,6 @@ func (s *DockerSuite) TestBuildHandleEscapes(c *check.C) {
 
 }
 
-func (s *DockerSuite) TestBuildOnBuildLowercase(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildonbuildlowercase"
-	name2 := "testbuildonbuildlowercase2"
-
-	_, err := buildImage(name,
-		`
-  FROM busybox
-  onbuild run echo quux
-  `, true)
-
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	_, out, err := buildImageWithOut(name2, fmt.Sprintf(`
-  FROM %s
-  `, name), true)
-
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	if !strings.Contains(out, "quux") {
-		c.Fatalf("Did not receive the expected echo text, got %s", out)
-	}
-
-	if strings.Contains(out, "ONBUILD ONBUILD") {
-		c.Fatalf("Got an ONBUILD ONBUILD error with no error: got %s", out)
-	}
-
-}
-
-func (s *DockerSuite) TestBuildEnvEscapes(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildenvescapes"
-	_, err := buildImage(name,
-		`
-    FROM busybox
-    ENV TEST foo
-    CMD echo \$
-    `,
-		true)
-
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	out, _ := dockerCmd(c, "run", "-t", name)
-
-	if strings.TrimSpace(out) != "$" {
-		c.Fatalf("Env TEST was not overwritten with bar when foo was supplied to dockerfile: was %q", strings.TrimSpace(out))
-	}
-
-}
-
-func (s *DockerSuite) TestBuildEnvOverwrite(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildenvoverwrite"
-
-	_, err := buildImage(name,
-		`
-    FROM busybox
-    ENV TEST foo
-    CMD echo ${TEST}
-    `,
-		true)
-
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	out, _ := dockerCmd(c, "run", "-e", "TEST=bar", "-t", name)
-
-	if strings.TrimSpace(out) != "bar" {
-		c.Fatalf("Env TEST was not overwritten with bar when foo was supplied to dockerfile: was %q", strings.TrimSpace(out))
-	}
-
-}
-
-func (s *DockerSuite) TestBuildOnBuildForbiddenMaintainerInSourceImage(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildonbuildforbiddenmaintainerinsourceimage"
-
-	out, _ := dockerCmd(c, "create", "busybox", "true")
-
-	cleanedContainerID := strings.TrimSpace(out)
-
-	dockerCmd(c, "commit", "--run", "{\"OnBuild\":[\"MAINTAINER docker.io\"]}", cleanedContainerID, "onbuild")
-
-	_, err := buildImage(name,
-		`FROM onbuild`,
-		true)
-	if err != nil {
-		if !strings.Contains(err.Error(), "maintainer isn't allowed as an ONBUILD trigger") {
-			c.Fatalf("Wrong error %v, must be about MAINTAINER and ONBUILD in source image", err)
-		}
-	} else {
-		c.Fatal("Error must not be nil")
-	}
-
-}
-
-func (s *DockerSuite) TestBuildOnBuildForbiddenFromInSourceImage(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildonbuildforbiddenfrominsourceimage"
-
-	out, _ := dockerCmd(c, "create", "busybox", "true")
-
-	cleanedContainerID := strings.TrimSpace(out)
-
-	dockerCmd(c, "commit", "--run", "{\"OnBuild\":[\"FROM busybox\"]}", cleanedContainerID, "onbuild")
-
-	_, err := buildImage(name,
-		`FROM onbuild`,
-		true)
-	if err != nil {
-		if !strings.Contains(err.Error(), "from isn't allowed as an ONBUILD trigger") {
-			c.Fatalf("Wrong error %v, must be about FROM and ONBUILD in source image", err)
-		}
-	} else {
-		c.Fatal("Error must not be nil")
-	}
-
-}
-
-func (s *DockerSuite) TestBuildOnBuildForbiddenChainedInSourceImage(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildonbuildforbiddenchainedinsourceimage"
-
-	out, _ := dockerCmd(c, "create", "busybox", "true")
-
-	cleanedContainerID := strings.TrimSpace(out)
-
-	dockerCmd(c, "commit", "--run", "{\"OnBuild\":[\"ONBUILD RUN ls\"]}", cleanedContainerID, "onbuild")
-
-	_, err := buildImage(name,
-		`FROM onbuild`,
-		true)
-	if err != nil {
-		if !strings.Contains(err.Error(), "Chaining ONBUILD via `ONBUILD ONBUILD` isn't allowed") {
-			c.Fatalf("Wrong error %v, must be about chaining ONBUILD in source image", err)
-		}
-	} else {
-		c.Fatal("Error must not be nil")
-	}
-
-}
-
-func (s *DockerSuite) TestBuildOnBuildCmdEntrypointJSON(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name1 := "onbuildcmd"
-	name2 := "onbuildgenerated"
-
-	_, err := buildImage(name1, `
-FROM busybox
-ONBUILD CMD ["hello world"]
-ONBUILD ENTRYPOINT ["echo"]
-ONBUILD RUN ["true"]`,
-		false)
-
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	_, err = buildImage(name2, fmt.Sprintf(`FROM %s`, name1), false)
-
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	out, _ := dockerCmd(c, "run", "-t", name2)
-
-	if !regexp.MustCompile(`(?m)^hello world`).MatchString(out) {
-		c.Fatal("did not get echo output from onbuild", out)
-	}
-
-}
-
-func (s *DockerSuite) TestBuildOnBuildEntrypointJSON(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name1 := "onbuildcmd"
-	name2 := "onbuildgenerated"
-
-	_, err := buildImage(name1, `
-FROM busybox
-ONBUILD ENTRYPOINT ["echo"]`,
-		false)
-
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	_, err = buildImage(name2, fmt.Sprintf("FROM %s\nCMD [\"hello world\"]\n", name1), false)
-
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	out, _ := dockerCmd(c, "run", "-t", name2)
-
-	if !regexp.MustCompile(`(?m)^hello world`).MatchString(out) {
-		c.Fatal("got malformed output from onbuild", out)
-	}
-
-}
-
-func (s *DockerSuite) TestBuildCacheAdd(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildtwoimageswithadd"
-	server, err := fakeStorage(map[string]string{
-		"robots.txt": "hello",
-		"index.html": "world",
-	})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer server.Close()
-
-	if _, err := buildImage(name,
-		fmt.Sprintf(`FROM scratch
-		ADD %s/robots.txt /`, server.URL()),
-		true); err != nil {
-		c.Fatal(err)
-	}
-	if err != nil {
-		c.Fatal(err)
-	}
-	deleteImages(name)
-	_, out, err := buildImageWithOut(name,
-		fmt.Sprintf(`FROM scratch
-		ADD %s/index.html /`, server.URL()),
-		true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	if strings.Contains(out, "Using cache") {
-		c.Fatal("2nd build used cache on ADD, it shouldn't")
-	}
-
-}
-
 func (s *DockerSuite) TestBuildLastModified(c *check.C) {
 	testRequires(c, DaemonIsLinux)
 	name := "testbuildlastmodified"
@@ -710,1131 +253,6 @@ func (s *DockerSuite) TestBuildSixtySteps(c *check.C) {
 
 	if _, err := buildImageFromContext(name, ctx, true); err != nil {
 		c.Fatal(err)
-	}
-}
-
-func (s *DockerSuite) TestBuildAddSingleFileToRoot(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testaddimg"
-	ctx, err := fakeContext(fmt.Sprintf(`FROM busybox
-RUN echo 'dockerio:x:1001:1001::/bin:/bin/false' >> /etc/passwd
-RUN echo 'dockerio:x:1001:' >> /etc/group
-RUN touch /exists
-RUN chown dockerio.dockerio /exists
-ADD test_file /
-RUN [ $(ls -l /test_file | awk '{print $3":"$4}') = 'root:root' ]
-RUN [ $(ls -l /test_file | awk '{print $1}') = '%s' ]
-RUN [ $(ls -l /exists | awk '{print $3":"$4}') = 'dockerio:dockerio' ]`, expectedFileChmod),
-		map[string]string{
-			"test_file": "test1",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-}
-
-// Issue #3960: "ADD src ." hangs
-func (s *DockerSuite) TestBuildAddSingleFileToWorkdir(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testaddsinglefiletoworkdir"
-	ctx, err := fakeContext(`FROM busybox
-ADD test_file .`,
-		map[string]string{
-			"test_file": "test1",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	errChan := make(chan error)
-	go func() {
-		_, err := buildImageFromContext(name, ctx, true)
-		errChan <- err
-		close(errChan)
-	}()
-	select {
-	case <-time.After(15 * time.Second):
-		c.Fatal("Build with adding to workdir timed out")
-	case err := <-errChan:
-		c.Assert(err, check.IsNil)
-	}
-}
-
-func (s *DockerSuite) TestBuildAddSingleFileToExistDir(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testaddsinglefiletoexistdir"
-	ctx, err := fakeContext(`FROM busybox
-RUN echo 'dockerio:x:1001:1001::/bin:/bin/false' >> /etc/passwd
-RUN echo 'dockerio:x:1001:' >> /etc/group
-RUN mkdir /exists
-RUN touch /exists/exists_file
-RUN chown -R dockerio.dockerio /exists
-ADD test_file /exists/
-RUN [ $(ls -l / | grep exists | awk '{print $3":"$4}') = 'dockerio:dockerio' ]
-RUN [ $(ls -l /exists/test_file | awk '{print $3":"$4}') = 'root:root' ]
-RUN [ $(ls -l /exists/exists_file | awk '{print $3":"$4}') = 'dockerio:dockerio' ]`,
-		map[string]string{
-			"test_file": "test1",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-}
-
-func (s *DockerSuite) TestBuildCopyAddMultipleFiles(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	server, err := fakeStorage(map[string]string{
-		"robots.txt": "hello",
-	})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer server.Close()
-
-	name := "testcopymultiplefilestofile"
-	ctx, err := fakeContext(fmt.Sprintf(`FROM busybox
-RUN echo 'dockerio:x:1001:1001::/bin:/bin/false' >> /etc/passwd
-RUN echo 'dockerio:x:1001:' >> /etc/group
-RUN mkdir /exists
-RUN touch /exists/exists_file
-RUN chown -R dockerio.dockerio /exists
-COPY test_file1 test_file2 /exists/
-ADD test_file3 test_file4 %s/robots.txt /exists/
-RUN [ $(ls -l / | grep exists | awk '{print $3":"$4}') = 'dockerio:dockerio' ]
-RUN [ $(ls -l /exists/test_file1 | awk '{print $3":"$4}') = 'root:root' ]
-RUN [ $(ls -l /exists/test_file2 | awk '{print $3":"$4}') = 'root:root' ]
-
-RUN [ $(ls -l /exists/test_file3 | awk '{print $3":"$4}') = 'root:root' ]
-RUN [ $(ls -l /exists/test_file4 | awk '{print $3":"$4}') = 'root:root' ]
-RUN [ $(ls -l /exists/robots.txt | awk '{print $3":"$4}') = 'root:root' ]
-
-RUN [ $(ls -l /exists/exists_file | awk '{print $3":"$4}') = 'dockerio:dockerio' ]
-`, server.URL()),
-		map[string]string{
-			"test_file1": "test1",
-			"test_file2": "test2",
-			"test_file3": "test3",
-			"test_file4": "test4",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-}
-
-func (s *DockerSuite) TestBuildAddMultipleFilesToFile(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testaddmultiplefilestofile"
-	ctx, err := fakeContext(`FROM scratch
-	ADD file1.txt file2.txt test
-	`,
-		map[string]string{
-			"file1.txt": "test1",
-			"file2.txt": "test1",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	expected := "When using ADD with more than one source file, the destination must be a directory and end with a /"
-	if _, err := buildImageFromContext(name, ctx, true); err == nil || !strings.Contains(err.Error(), expected) {
-		c.Fatalf("Wrong error: (should contain %q) got:\n%v", expected, err)
-	}
-
-}
-
-func (s *DockerSuite) TestBuildJSONAddMultipleFilesToFile(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testjsonaddmultiplefilestofile"
-	ctx, err := fakeContext(`FROM scratch
-	ADD ["file1.txt", "file2.txt", "test"]
-	`,
-		map[string]string{
-			"file1.txt": "test1",
-			"file2.txt": "test1",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	expected := "When using ADD with more than one source file, the destination must be a directory and end with a /"
-	if _, err := buildImageFromContext(name, ctx, true); err == nil || !strings.Contains(err.Error(), expected) {
-		c.Fatalf("Wrong error: (should contain %q) got:\n%v", expected, err)
-	}
-
-}
-
-func (s *DockerSuite) TestBuildAddMultipleFilesToFileWild(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testaddmultiplefilestofilewild"
-	ctx, err := fakeContext(`FROM scratch
-	ADD file*.txt test
-	`,
-		map[string]string{
-			"file1.txt": "test1",
-			"file2.txt": "test1",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	expected := "When using ADD with more than one source file, the destination must be a directory and end with a /"
-	if _, err := buildImageFromContext(name, ctx, true); err == nil || !strings.Contains(err.Error(), expected) {
-		c.Fatalf("Wrong error: (should contain %q) got:\n%v", expected, err)
-	}
-
-}
-
-func (s *DockerSuite) TestBuildJSONAddMultipleFilesToFileWild(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testjsonaddmultiplefilestofilewild"
-	ctx, err := fakeContext(`FROM scratch
-	ADD ["file*.txt", "test"]
-	`,
-		map[string]string{
-			"file1.txt": "test1",
-			"file2.txt": "test1",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	expected := "When using ADD with more than one source file, the destination must be a directory and end with a /"
-	if _, err := buildImageFromContext(name, ctx, true); err == nil || !strings.Contains(err.Error(), expected) {
-		c.Fatalf("Wrong error: (should contain %q) got:\n%v", expected, err)
-	}
-
-}
-
-func (s *DockerSuite) TestBuildCopyMultipleFilesToFile(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testcopymultiplefilestofile"
-	ctx, err := fakeContext(`FROM scratch
-	COPY file1.txt file2.txt test
-	`,
-		map[string]string{
-			"file1.txt": "test1",
-			"file2.txt": "test1",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	expected := "When using COPY with more than one source file, the destination must be a directory and end with a /"
-	if _, err := buildImageFromContext(name, ctx, true); err == nil || !strings.Contains(err.Error(), expected) {
-		c.Fatalf("Wrong error: (should contain %q) got:\n%v", expected, err)
-	}
-
-}
-
-func (s *DockerSuite) TestBuildJSONCopyMultipleFilesToFile(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testjsoncopymultiplefilestofile"
-	ctx, err := fakeContext(`FROM scratch
-	COPY ["file1.txt", "file2.txt", "test"]
-	`,
-		map[string]string{
-			"file1.txt": "test1",
-			"file2.txt": "test1",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	expected := "When using COPY with more than one source file, the destination must be a directory and end with a /"
-	if _, err := buildImageFromContext(name, ctx, true); err == nil || !strings.Contains(err.Error(), expected) {
-		c.Fatalf("Wrong error: (should contain %q) got:\n%v", expected, err)
-	}
-
-}
-
-func (s *DockerSuite) TestBuildAddFileWithWhitespace(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testaddfilewithwhitespace"
-	ctx, err := fakeContext(`FROM busybox
-RUN mkdir "/test dir"
-RUN mkdir "/test_dir"
-ADD [ "test file1", "/test_file1" ]
-ADD [ "test_file2", "/test file2" ]
-ADD [ "test file3", "/test file3" ]
-ADD [ "test dir/test_file4", "/test_dir/test_file4" ]
-ADD [ "test_dir/test_file5", "/test dir/test_file5" ]
-ADD [ "test dir/test_file6", "/test dir/test_file6" ]
-RUN [ $(cat "/test_file1") = 'test1' ]
-RUN [ $(cat "/test file2") = 'test2' ]
-RUN [ $(cat "/test file3") = 'test3' ]
-RUN [ $(cat "/test_dir/test_file4") = 'test4' ]
-RUN [ $(cat "/test dir/test_file5") = 'test5' ]
-RUN [ $(cat "/test dir/test_file6") = 'test6' ]`,
-		map[string]string{
-			"test file1":          "test1",
-			"test_file2":          "test2",
-			"test file3":          "test3",
-			"test dir/test_file4": "test4",
-			"test_dir/test_file5": "test5",
-			"test dir/test_file6": "test6",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-}
-
-func (s *DockerSuite) TestBuildCopyFileWithWhitespace(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testcopyfilewithwhitespace"
-	ctx, err := fakeContext(`FROM busybox
-RUN mkdir "/test dir"
-RUN mkdir "/test_dir"
-COPY [ "test file1", "/test_file1" ]
-COPY [ "test_file2", "/test file2" ]
-COPY [ "test file3", "/test file3" ]
-COPY [ "test dir/test_file4", "/test_dir/test_file4" ]
-COPY [ "test_dir/test_file5", "/test dir/test_file5" ]
-COPY [ "test dir/test_file6", "/test dir/test_file6" ]
-RUN [ $(cat "/test_file1") = 'test1' ]
-RUN [ $(cat "/test file2") = 'test2' ]
-RUN [ $(cat "/test file3") = 'test3' ]
-RUN [ $(cat "/test_dir/test_file4") = 'test4' ]
-RUN [ $(cat "/test dir/test_file5") = 'test5' ]
-RUN [ $(cat "/test dir/test_file6") = 'test6' ]`,
-		map[string]string{
-			"test file1":          "test1",
-			"test_file2":          "test2",
-			"test file3":          "test3",
-			"test dir/test_file4": "test4",
-			"test_dir/test_file5": "test5",
-			"test dir/test_file6": "test6",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-}
-
-func (s *DockerSuite) TestBuildAddMultipleFilesToFileWithWhitespace(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testaddmultiplefilestofilewithwhitespace"
-	ctx, err := fakeContext(`FROM busybox
-	ADD [ "test file1", "test file2", "test" ]
-    `,
-		map[string]string{
-			"test file1": "test1",
-			"test file2": "test2",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	expected := "When using ADD with more than one source file, the destination must be a directory and end with a /"
-	if _, err := buildImageFromContext(name, ctx, true); err == nil || !strings.Contains(err.Error(), expected) {
-		c.Fatalf("Wrong error: (should contain %q) got:\n%v", expected, err)
-	}
-
-}
-
-func (s *DockerSuite) TestBuildCopyMultipleFilesToFileWithWhitespace(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testcopymultiplefilestofilewithwhitespace"
-	ctx, err := fakeContext(`FROM busybox
-	COPY [ "test file1", "test file2", "test" ]
-        `,
-		map[string]string{
-			"test file1": "test1",
-			"test file2": "test2",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	expected := "When using COPY with more than one source file, the destination must be a directory and end with a /"
-	if _, err := buildImageFromContext(name, ctx, true); err == nil || !strings.Contains(err.Error(), expected) {
-		c.Fatalf("Wrong error: (should contain %q) got:\n%v", expected, err)
-	}
-
-}
-
-func (s *DockerSuite) TestBuildCopyWildcard(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testcopywildcard"
-	server, err := fakeStorage(map[string]string{
-		"robots.txt": "hello",
-		"index.html": "world",
-	})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer server.Close()
-
-	ctx, err := fakeContext(fmt.Sprintf(`FROM busybox
-	COPY file*.txt /tmp/
-	RUN ls /tmp/file1.txt /tmp/file2.txt
-	RUN mkdir /tmp1
-	COPY dir* /tmp1/
-	RUN ls /tmp1/dirt /tmp1/nested_file /tmp1/nested_dir/nest_nest_file
-	RUN mkdir /tmp2
-        ADD dir/*dir %s/robots.txt /tmp2/
-	RUN ls /tmp2/nest_nest_file /tmp2/robots.txt
-	`, server.URL()),
-		map[string]string{
-			"file1.txt":                     "test1",
-			"file2.txt":                     "test2",
-			"dir/nested_file":               "nested file",
-			"dir/nested_dir/nest_nest_file": "2 times nested",
-			"dirt": "dirty",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	id1, err := buildImageFromContext(name, ctx, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	// Now make sure we use a cache the 2nd time
-	id2, err := buildImageFromContext(name, ctx, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	if id1 != id2 {
-		c.Fatal("didn't use the cache")
-	}
-
-}
-
-func (s *DockerSuite) TestBuildCopyWildcardNoFind(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testcopywildcardnofind"
-	ctx, err := fakeContext(`FROM busybox
-	COPY file*.txt /tmp/
-	`, nil)
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	_, err = buildImageFromContext(name, ctx, true)
-	if err == nil {
-		c.Fatal("should have failed to find a file")
-	}
-	if !strings.Contains(err.Error(), "No source files were specified") {
-		c.Fatalf("Wrong error %v, must be about no source files", err)
-	}
-
-}
-
-func (s *DockerSuite) TestBuildCopyWildcardInName(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testcopywildcardinname"
-	ctx, err := fakeContext(`FROM busybox
-	COPY *.txt /tmp/
-	RUN [ "$(cat /tmp/\*.txt)" = 'hi there' ]
-	`, map[string]string{"*.txt": "hi there"})
-
-	if err != nil {
-		// Normally we would do c.Fatal(err) here but given that
-		// the odds of this failing are so rare, it must be because
-		// the OS we're running the client on doesn't support * in
-		// filenames (like windows).  So, instead of failing the test
-		// just let it pass. Then we don't need to explicitly
-		// say which OSs this works on or not.
-		return
-	}
-	defer ctx.Close()
-
-	_, err = buildImageFromContext(name, ctx, true)
-	if err != nil {
-		c.Fatalf("should have built: %q", err)
-	}
-}
-
-func (s *DockerSuite) TestBuildCopyWildcardCache(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testcopywildcardcache"
-	ctx, err := fakeContext(`FROM busybox
-	COPY file1.txt /tmp/`,
-		map[string]string{
-			"file1.txt": "test1",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	id1, err := buildImageFromContext(name, ctx, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	// Now make sure we use a cache the 2nd time even with wild cards.
-	// Use the same context so the file is the same and the checksum will match
-	ctx.Add("Dockerfile", `FROM busybox
-	COPY file*.txt /tmp/`)
-
-	id2, err := buildImageFromContext(name, ctx, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	if id1 != id2 {
-		c.Fatal("didn't use the cache")
-	}
-
-}
-
-func (s *DockerSuite) TestBuildAddSingleFileToNonExistingDir(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testaddsinglefiletononexistingdir"
-	ctx, err := fakeContext(`FROM busybox
-RUN echo 'dockerio:x:1001:1001::/bin:/bin/false' >> /etc/passwd
-RUN echo 'dockerio:x:1001:' >> /etc/group
-RUN touch /exists
-RUN chown dockerio.dockerio /exists
-ADD test_file /test_dir/
-RUN [ $(ls -l / | grep test_dir | awk '{print $3":"$4}') = 'root:root' ]
-RUN [ $(ls -l /test_dir/test_file | awk '{print $3":"$4}') = 'root:root' ]
-RUN [ $(ls -l /exists | awk '{print $3":"$4}') = 'dockerio:dockerio' ]`,
-		map[string]string{
-			"test_file": "test1",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-
-}
-
-func (s *DockerSuite) TestBuildAddDirContentToRoot(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testadddircontenttoroot"
-	ctx, err := fakeContext(`FROM busybox
-RUN echo 'dockerio:x:1001:1001::/bin:/bin/false' >> /etc/passwd
-RUN echo 'dockerio:x:1001:' >> /etc/group
-RUN touch /exists
-RUN chown dockerio.dockerio exists
-ADD test_dir /
-RUN [ $(ls -l /test_file | awk '{print $3":"$4}') = 'root:root' ]
-RUN [ $(ls -l /exists | awk '{print $3":"$4}') = 'dockerio:dockerio' ]`,
-		map[string]string{
-			"test_dir/test_file": "test1",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-}
-
-func (s *DockerSuite) TestBuildAddDirContentToExistingDir(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testadddircontenttoexistingdir"
-	ctx, err := fakeContext(`FROM busybox
-RUN echo 'dockerio:x:1001:1001::/bin:/bin/false' >> /etc/passwd
-RUN echo 'dockerio:x:1001:' >> /etc/group
-RUN mkdir /exists
-RUN touch /exists/exists_file
-RUN chown -R dockerio.dockerio /exists
-ADD test_dir/ /exists/
-RUN [ $(ls -l / | grep exists | awk '{print $3":"$4}') = 'dockerio:dockerio' ]
-RUN [ $(ls -l /exists/exists_file | awk '{print $3":"$4}') = 'dockerio:dockerio' ]
-RUN [ $(ls -l /exists/test_file | awk '{print $3":"$4}') = 'root:root' ]`,
-		map[string]string{
-			"test_dir/test_file": "test1",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-}
-
-func (s *DockerSuite) TestBuildAddWholeDirToRoot(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testaddwholedirtoroot"
-	ctx, err := fakeContext(fmt.Sprintf(`FROM busybox
-RUN echo 'dockerio:x:1001:1001::/bin:/bin/false' >> /etc/passwd
-RUN echo 'dockerio:x:1001:' >> /etc/group
-RUN touch /exists
-RUN chown dockerio.dockerio exists
-ADD test_dir /test_dir
-RUN [ $(ls -l / | grep test_dir | awk '{print $3":"$4}') = 'root:root' ]
-RUN [ $(ls -l / | grep test_dir | awk '{print $1}') = 'drwxr-xr-x' ]
-RUN [ $(ls -l /test_dir/test_file | awk '{print $3":"$4}') = 'root:root' ]
-RUN [ $(ls -l /test_dir/test_file | awk '{print $1}') = '%s' ]
-RUN [ $(ls -l /exists | awk '{print $3":"$4}') = 'dockerio:dockerio' ]`, expectedFileChmod),
-		map[string]string{
-			"test_dir/test_file": "test1",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-}
-
-// Testing #5941
-func (s *DockerSuite) TestBuildAddEtcToRoot(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testaddetctoroot"
-	ctx, err := fakeContext(`FROM scratch
-ADD . /`,
-		map[string]string{
-			"etc/test_file": "test1",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-}
-
-// Testing #9401
-func (s *DockerSuite) TestBuildAddPreservesFilesSpecialBits(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testaddpreservesfilesspecialbits"
-	ctx, err := fakeContext(`FROM busybox
-ADD suidbin /usr/bin/suidbin
-RUN chmod 4755 /usr/bin/suidbin
-RUN [ $(ls -l /usr/bin/suidbin | awk '{print $1}') = '-rwsr-xr-x' ]
-ADD ./data/ /
-RUN [ $(ls -l /usr/bin/suidbin | awk '{print $1}') = '-rwsr-xr-x' ]`,
-		map[string]string{
-			"suidbin":             "suidbin",
-			"/data/usr/test_file": "test1",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-}
-
-func (s *DockerSuite) TestBuildCopySingleFileToRoot(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testcopysinglefiletoroot"
-	ctx, err := fakeContext(fmt.Sprintf(`FROM busybox
-RUN echo 'dockerio:x:1001:1001::/bin:/bin/false' >> /etc/passwd
-RUN echo 'dockerio:x:1001:' >> /etc/group
-RUN touch /exists
-RUN chown dockerio.dockerio /exists
-COPY test_file /
-RUN [ $(ls -l /test_file | awk '{print $3":"$4}') = 'root:root' ]
-RUN [ $(ls -l /test_file | awk '{print $1}') = '%s' ]
-RUN [ $(ls -l /exists | awk '{print $3":"$4}') = 'dockerio:dockerio' ]`, expectedFileChmod),
-		map[string]string{
-			"test_file": "test1",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-}
-
-// Issue #3960: "ADD src ." hangs - adapted for COPY
-func (s *DockerSuite) TestBuildCopySingleFileToWorkdir(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testcopysinglefiletoworkdir"
-	ctx, err := fakeContext(`FROM busybox
-COPY test_file .`,
-		map[string]string{
-			"test_file": "test1",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	errChan := make(chan error)
-	go func() {
-		_, err := buildImageFromContext(name, ctx, true)
-		errChan <- err
-		close(errChan)
-	}()
-	select {
-	case <-time.After(15 * time.Second):
-		c.Fatal("Build with adding to workdir timed out")
-	case err := <-errChan:
-		c.Assert(err, check.IsNil)
-	}
-}
-
-func (s *DockerSuite) TestBuildCopySingleFileToExistDir(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testcopysinglefiletoexistdir"
-	ctx, err := fakeContext(`FROM busybox
-RUN echo 'dockerio:x:1001:1001::/bin:/bin/false' >> /etc/passwd
-RUN echo 'dockerio:x:1001:' >> /etc/group
-RUN mkdir /exists
-RUN touch /exists/exists_file
-RUN chown -R dockerio.dockerio /exists
-COPY test_file /exists/
-RUN [ $(ls -l / | grep exists | awk '{print $3":"$4}') = 'dockerio:dockerio' ]
-RUN [ $(ls -l /exists/test_file | awk '{print $3":"$4}') = 'root:root' ]
-RUN [ $(ls -l /exists/exists_file | awk '{print $3":"$4}') = 'dockerio:dockerio' ]`,
-		map[string]string{
-			"test_file": "test1",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-}
-
-func (s *DockerSuite) TestBuildCopySingleFileToNonExistDir(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testcopysinglefiletononexistdir"
-	ctx, err := fakeContext(`FROM busybox
-RUN echo 'dockerio:x:1001:1001::/bin:/bin/false' >> /etc/passwd
-RUN echo 'dockerio:x:1001:' >> /etc/group
-RUN touch /exists
-RUN chown dockerio.dockerio /exists
-COPY test_file /test_dir/
-RUN [ $(ls -l / | grep test_dir | awk '{print $3":"$4}') = 'root:root' ]
-RUN [ $(ls -l /test_dir/test_file | awk '{print $3":"$4}') = 'root:root' ]
-RUN [ $(ls -l /exists | awk '{print $3":"$4}') = 'dockerio:dockerio' ]`,
-		map[string]string{
-			"test_file": "test1",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-}
-
-func (s *DockerSuite) TestBuildCopyDirContentToRoot(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testcopydircontenttoroot"
-	ctx, err := fakeContext(`FROM busybox
-RUN echo 'dockerio:x:1001:1001::/bin:/bin/false' >> /etc/passwd
-RUN echo 'dockerio:x:1001:' >> /etc/group
-RUN touch /exists
-RUN chown dockerio.dockerio exists
-COPY test_dir /
-RUN [ $(ls -l /test_file | awk '{print $3":"$4}') = 'root:root' ]
-RUN [ $(ls -l /exists | awk '{print $3":"$4}') = 'dockerio:dockerio' ]`,
-		map[string]string{
-			"test_dir/test_file": "test1",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-}
-
-func (s *DockerSuite) TestBuildCopyDirContentToExistDir(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testcopydircontenttoexistdir"
-	ctx, err := fakeContext(`FROM busybox
-RUN echo 'dockerio:x:1001:1001::/bin:/bin/false' >> /etc/passwd
-RUN echo 'dockerio:x:1001:' >> /etc/group
-RUN mkdir /exists
-RUN touch /exists/exists_file
-RUN chown -R dockerio.dockerio /exists
-COPY test_dir/ /exists/
-RUN [ $(ls -l / | grep exists | awk '{print $3":"$4}') = 'dockerio:dockerio' ]
-RUN [ $(ls -l /exists/exists_file | awk '{print $3":"$4}') = 'dockerio:dockerio' ]
-RUN [ $(ls -l /exists/test_file | awk '{print $3":"$4}') = 'root:root' ]`,
-		map[string]string{
-			"test_dir/test_file": "test1",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-}
-
-func (s *DockerSuite) TestBuildCopyWholeDirToRoot(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testcopywholedirtoroot"
-	ctx, err := fakeContext(fmt.Sprintf(`FROM busybox
-RUN echo 'dockerio:x:1001:1001::/bin:/bin/false' >> /etc/passwd
-RUN echo 'dockerio:x:1001:' >> /etc/group
-RUN touch /exists
-RUN chown dockerio.dockerio exists
-COPY test_dir /test_dir
-RUN [ $(ls -l / | grep test_dir | awk '{print $3":"$4}') = 'root:root' ]
-RUN [ $(ls -l / | grep test_dir | awk '{print $1}') = 'drwxr-xr-x' ]
-RUN [ $(ls -l /test_dir/test_file | awk '{print $3":"$4}') = 'root:root' ]
-RUN [ $(ls -l /test_dir/test_file | awk '{print $1}') = '%s' ]
-RUN [ $(ls -l /exists | awk '{print $3":"$4}') = 'dockerio:dockerio' ]`, expectedFileChmod),
-		map[string]string{
-			"test_dir/test_file": "test1",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-}
-
-func (s *DockerSuite) TestBuildCopyEtcToRoot(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testcopyetctoroot"
-	ctx, err := fakeContext(`FROM scratch
-COPY . /`,
-		map[string]string{
-			"etc/test_file": "test1",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-}
-
-func (s *DockerSuite) TestBuildCopyDisallowRemote(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testcopydisallowremote"
-	_, out, err := buildImageWithOut(name, `FROM scratch
-COPY https://index.docker.io/robots.txt /`,
-		true)
-	if err == nil || !strings.Contains(out, "Source can't be a URL for COPY") {
-		c.Fatalf("Error should be about disallowed remote source, got err: %s, out: %q", err, out)
-	}
-}
-
-func (s *DockerSuite) TestBuildAddBadLinks(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	const (
-		dockerfile = `
-			FROM scratch
-			ADD links.tar /
-			ADD foo.txt /symlink/
-			`
-		targetFile = "foo.txt"
-	)
-	var (
-		name = "test-link-absolute"
-	)
-	ctx, err := fakeContext(dockerfile, nil)
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	tempDir, err := ioutil.TempDir("", "test-link-absolute-temp-")
-	if err != nil {
-		c.Fatalf("failed to create temporary directory: %s", tempDir)
-	}
-	defer os.RemoveAll(tempDir)
-
-	var symlinkTarget string
-	if runtime.GOOS == "windows" {
-		var driveLetter string
-		if abs, err := filepath.Abs(tempDir); err != nil {
-			c.Fatal(err)
-		} else {
-			driveLetter = abs[:1]
-		}
-		tempDirWithoutDrive := tempDir[2:]
-		symlinkTarget = fmt.Sprintf(`%s:\..\..\..\..\..\..\..\..\..\..\..\..%s`, driveLetter, tempDirWithoutDrive)
-	} else {
-		symlinkTarget = fmt.Sprintf("/../../../../../../../../../../../..%s", tempDir)
-	}
-
-	tarPath := filepath.Join(ctx.Dir, "links.tar")
-	nonExistingFile := filepath.Join(tempDir, targetFile)
-	fooPath := filepath.Join(ctx.Dir, targetFile)
-
-	tarOut, err := os.Create(tarPath)
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	tarWriter := tar.NewWriter(tarOut)
-
-	header := &tar.Header{
-		Name:     "symlink",
-		Typeflag: tar.TypeSymlink,
-		Linkname: symlinkTarget,
-		Mode:     0755,
-		Uid:      0,
-		Gid:      0,
-	}
-
-	err = tarWriter.WriteHeader(header)
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	tarWriter.Close()
-	tarOut.Close()
-
-	foo, err := os.Create(fooPath)
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer foo.Close()
-
-	if _, err := foo.WriteString("test"); err != nil {
-		c.Fatal(err)
-	}
-
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-
-	if _, err := os.Stat(nonExistingFile); err == nil || err != nil && !os.IsNotExist(err) {
-		c.Fatalf("%s shouldn't have been written and it shouldn't exist", nonExistingFile)
-	}
-
-}
-
-func (s *DockerSuite) TestBuildAddBadLinksVolume(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	const (
-		dockerfileTemplate = `
-		FROM busybox
-		RUN ln -s /../../../../../../../../%s /x
-		VOLUME /x
-		ADD foo.txt /x/`
-		targetFile = "foo.txt"
-	)
-	var (
-		name       = "test-link-absolute-volume"
-		dockerfile = ""
-	)
-
-	tempDir, err := ioutil.TempDir("", "test-link-absolute-volume-temp-")
-	if err != nil {
-		c.Fatalf("failed to create temporary directory: %s", tempDir)
-	}
-	defer os.RemoveAll(tempDir)
-
-	dockerfile = fmt.Sprintf(dockerfileTemplate, tempDir)
-	nonExistingFile := filepath.Join(tempDir, targetFile)
-
-	ctx, err := fakeContext(dockerfile, nil)
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-	fooPath := filepath.Join(ctx.Dir, targetFile)
-
-	foo, err := os.Create(fooPath)
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer foo.Close()
-
-	if _, err := foo.WriteString("test"); err != nil {
-		c.Fatal(err)
-	}
-
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-
-	if _, err := os.Stat(nonExistingFile); err == nil || err != nil && !os.IsNotExist(err) {
-		c.Fatalf("%s shouldn't have been written and it shouldn't exist", nonExistingFile)
-	}
-
-}
-
-// Issue #5270 - ensure we throw a better error than "unexpected EOF"
-// when we can't access files in the context.
-func (s *DockerSuite) TestBuildWithInaccessibleFilesInContext(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	testRequires(c, UnixCli) // test uses chown/chmod: not available on windows
-
-	{
-		name := "testbuildinaccessiblefiles"
-		ctx, err := fakeContext("FROM scratch\nADD . /foo/", map[string]string{"fileWithoutReadAccess": "foo"})
-		if err != nil {
-			c.Fatal(err)
-		}
-		defer ctx.Close()
-		// This is used to ensure we detect inaccessible files early during build in the cli client
-		pathToFileWithoutReadAccess := filepath.Join(ctx.Dir, "fileWithoutReadAccess")
-
-		if err = os.Chown(pathToFileWithoutReadAccess, 0, 0); err != nil {
-			c.Fatalf("failed to chown file to root: %s", err)
-		}
-		if err = os.Chmod(pathToFileWithoutReadAccess, 0700); err != nil {
-			c.Fatalf("failed to chmod file to 700: %s", err)
-		}
-		buildCmd := exec.Command("su", "unprivilegeduser", "-c", fmt.Sprintf("%s build -t %s .", dockerBinary, name))
-		buildCmd.Dir = ctx.Dir
-		out, _, err := runCommandWithOutput(buildCmd)
-		if err == nil {
-			c.Fatalf("build should have failed: %s %s", err, out)
-		}
-
-		// check if we've detected the failure before we started building
-		if !strings.Contains(out, "no permission to read from ") {
-			c.Fatalf("output should've contained the string: no permission to read from but contained: %s", out)
-		}
-
-		if !strings.Contains(out, "Error checking context") {
-			c.Fatalf("output should've contained the string: Error checking context")
-		}
-	}
-	{
-		name := "testbuildinaccessibledirectory"
-		ctx, err := fakeContext("FROM scratch\nADD . /foo/", map[string]string{"directoryWeCantStat/bar": "foo"})
-		if err != nil {
-			c.Fatal(err)
-		}
-		defer ctx.Close()
-		// This is used to ensure we detect inaccessible directories early during build in the cli client
-		pathToDirectoryWithoutReadAccess := filepath.Join(ctx.Dir, "directoryWeCantStat")
-		pathToFileInDirectoryWithoutReadAccess := filepath.Join(pathToDirectoryWithoutReadAccess, "bar")
-
-		if err = os.Chown(pathToDirectoryWithoutReadAccess, 0, 0); err != nil {
-			c.Fatalf("failed to chown directory to root: %s", err)
-		}
-		if err = os.Chmod(pathToDirectoryWithoutReadAccess, 0444); err != nil {
-			c.Fatalf("failed to chmod directory to 444: %s", err)
-		}
-		if err = os.Chmod(pathToFileInDirectoryWithoutReadAccess, 0700); err != nil {
-			c.Fatalf("failed to chmod file to 700: %s", err)
-		}
-
-		buildCmd := exec.Command("su", "unprivilegeduser", "-c", fmt.Sprintf("%s build -t %s .", dockerBinary, name))
-		buildCmd.Dir = ctx.Dir
-		out, _, err := runCommandWithOutput(buildCmd)
-		if err == nil {
-			c.Fatalf("build should have failed: %s %s", err, out)
-		}
-
-		// check if we've detected the failure before we started building
-		if !strings.Contains(out, "can't stat") {
-			c.Fatalf("output should've contained the string: can't access %s", out)
-		}
-
-		if !strings.Contains(out, "Error checking context") {
-			c.Fatalf("output should've contained the string: Error checking context\ngot:%s", out)
-		}
-
-	}
-	{
-		name := "testlinksok"
-		ctx, err := fakeContext("FROM scratch\nADD . /foo/", nil)
-		if err != nil {
-			c.Fatal(err)
-		}
-		defer ctx.Close()
-
-		target := "../../../../../../../../../../../../../../../../../../../azA"
-		if err := os.Symlink(filepath.Join(ctx.Dir, "g"), target); err != nil {
-			c.Fatal(err)
-		}
-		defer os.Remove(target)
-		// This is used to ensure we don't follow links when checking if everything in the context is accessible
-		// This test doesn't require that we run commands as an unprivileged user
-		if _, err := buildImageFromContext(name, ctx, true); err != nil {
-			c.Fatal(err)
-		}
-	}
-	{
-		name := "testbuildignoredinaccessible"
-		ctx, err := fakeContext("FROM scratch\nADD . /foo/",
-			map[string]string{
-				"directoryWeCantStat/bar": "foo",
-				".dockerignore":           "directoryWeCantStat",
-			})
-		if err != nil {
-			c.Fatal(err)
-		}
-		defer ctx.Close()
-		// This is used to ensure we don't try to add inaccessible files when they are ignored by a .dockerignore pattern
-		pathToDirectoryWithoutReadAccess := filepath.Join(ctx.Dir, "directoryWeCantStat")
-		pathToFileInDirectoryWithoutReadAccess := filepath.Join(pathToDirectoryWithoutReadAccess, "bar")
-		if err = os.Chown(pathToDirectoryWithoutReadAccess, 0, 0); err != nil {
-			c.Fatalf("failed to chown directory to root: %s", err)
-		}
-		if err = os.Chmod(pathToDirectoryWithoutReadAccess, 0444); err != nil {
-			c.Fatalf("failed to chmod directory to 755: %s", err)
-		}
-		if err = os.Chmod(pathToFileInDirectoryWithoutReadAccess, 0700); err != nil {
-			c.Fatalf("failed to chmod file to 444: %s", err)
-		}
-
-		buildCmd := exec.Command("su", "unprivilegeduser", "-c", fmt.Sprintf("%s build -t %s .", dockerBinary, name))
-		buildCmd.Dir = ctx.Dir
-		if out, _, err := runCommandWithOutput(buildCmd); err != nil {
-			c.Fatalf("build should have worked: %s %s", err, out)
-		}
-
 	}
 }
 
@@ -2166,144 +584,6 @@ func (s *DockerSuite) TestBuildRelativeWorkdir(c *check.C) {
 	}
 }
 
-func (s *DockerSuite) TestBuildWorkdirWithEnvVariables(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildworkdirwithenvvariables"
-	expected := "/test1/test2"
-	_, err := buildImage(name,
-		`FROM busybox
-		ENV DIRPATH /test1
-		ENV SUBDIRNAME test2
-		WORKDIR $DIRPATH
-		WORKDIR $SUBDIRNAME/$MISSING_VAR`,
-		true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	res, err := inspectField(name, "Config.WorkingDir")
-	if err != nil {
-		c.Fatal(err)
-	}
-	if res != expected {
-		c.Fatalf("Workdir %s, expected %s", res, expected)
-	}
-}
-
-func (s *DockerSuite) TestBuildRelativeCopy(c *check.C) {
-	// cat /test1/test2/foo gets permission denied for the user
-	testRequires(c, NotUserNamespace)
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildrelativecopy"
-	dockerfile := `
-		FROM busybox
-			WORKDIR /test1
-			WORKDIR test2
-			RUN [ "$PWD" = '/test1/test2' ]
-			COPY foo ./
-			RUN [ "$(cat /test1/test2/foo)" = 'hello' ]
-			ADD foo ./bar/baz
-			RUN [ "$(cat /test1/test2/bar/baz)" = 'hello' ]
-			COPY foo ./bar/baz2
-			RUN [ "$(cat /test1/test2/bar/baz2)" = 'hello' ]
-			WORKDIR ..
-			COPY foo ./
-			RUN [ "$(cat /test1/foo)" = 'hello' ]
-			COPY foo /test3/
-			RUN [ "$(cat /test3/foo)" = 'hello' ]
-			WORKDIR /test4
-			COPY . .
-			RUN [ "$(cat /test4/foo)" = 'hello' ]
-			WORKDIR /test5/test6
-			COPY foo ../
-			RUN [ "$(cat /test5/foo)" = 'hello' ]
-			`
-	ctx, err := fakeContext(dockerfile, map[string]string{
-		"foo": "hello",
-	})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-	_, err = buildImageFromContext(name, ctx, false)
-	if err != nil {
-		c.Fatal(err)
-	}
-}
-
-func (s *DockerSuite) TestBuildEnv(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildenv"
-	expected := "[PATH=/test:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin PORT=2375]"
-	_, err := buildImage(name,
-		`FROM busybox
-		ENV PATH /test:$PATH
-        ENV PORT 2375
-		RUN [ $(env | grep PORT) = 'PORT=2375' ]`,
-		true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	res, err := inspectField(name, "Config.Env")
-	if err != nil {
-		c.Fatal(err)
-	}
-	if res != expected {
-		c.Fatalf("Env %s, expected %s", res, expected)
-	}
-}
-
-func (s *DockerSuite) TestBuildContextCleanup(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	testRequires(c, SameHostDaemon)
-
-	name := "testbuildcontextcleanup"
-	entries, err := ioutil.ReadDir(filepath.Join(dockerBasePath, "tmp"))
-	if err != nil {
-		c.Fatalf("failed to list contents of tmp dir: %s", err)
-	}
-	_, err = buildImage(name,
-		`FROM scratch
-        ENTRYPOINT ["/bin/echo"]`,
-		true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	entriesFinal, err := ioutil.ReadDir(filepath.Join(dockerBasePath, "tmp"))
-	if err != nil {
-		c.Fatalf("failed to list contents of tmp dir: %s", err)
-	}
-	if err = compareDirectoryEntries(entries, entriesFinal); err != nil {
-		c.Fatalf("context should have been deleted, but wasn't")
-	}
-
-}
-
-func (s *DockerSuite) TestBuildContextCleanupFailedBuild(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	testRequires(c, SameHostDaemon)
-
-	name := "testbuildcontextcleanup"
-	entries, err := ioutil.ReadDir(filepath.Join(dockerBasePath, "tmp"))
-	if err != nil {
-		c.Fatalf("failed to list contents of tmp dir: %s", err)
-	}
-	_, err = buildImage(name,
-		`FROM scratch
-	RUN /non/existing/command`,
-		true)
-	if err == nil {
-		c.Fatalf("expected build to fail, but it didn't")
-	}
-	entriesFinal, err := ioutil.ReadDir(filepath.Join(dockerBasePath, "tmp"))
-	if err != nil {
-		c.Fatalf("failed to list contents of tmp dir: %s", err)
-	}
-	if err = compareDirectoryEntries(entries, entriesFinal); err != nil {
-		c.Fatalf("context should have been deleted, but wasn't")
-	}
-
-}
-
 func (s *DockerSuite) TestBuildCmd(c *check.C) {
 	testRequires(c, DaemonIsLinux)
 	name := "testbuildcmd"
@@ -2321,123 +601,6 @@ func (s *DockerSuite) TestBuildCmd(c *check.C) {
 	}
 	if res != expected {
 		c.Fatalf("Cmd %s, expected %s", res, expected)
-	}
-}
-
-func (s *DockerSuite) TestBuildExpose(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildexpose"
-	expected := "map[2375/tcp:{}]"
-	_, err := buildImage(name,
-		`FROM scratch
-        EXPOSE 2375`,
-		true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	res, err := inspectField(name, "Config.ExposedPorts")
-	if err != nil {
-		c.Fatal(err)
-	}
-	if res != expected {
-		c.Fatalf("Exposed ports %s, expected %s", res, expected)
-	}
-}
-
-func (s *DockerSuite) TestBuildExposeMorePorts(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	// start building docker file with a large number of ports
-	portList := make([]string, 50)
-	line := make([]string, 100)
-	expectedPorts := make([]int, len(portList)*len(line))
-	for i := 0; i < len(portList); i++ {
-		for j := 0; j < len(line); j++ {
-			p := i*len(line) + j + 1
-			line[j] = strconv.Itoa(p)
-			expectedPorts[p-1] = p
-		}
-		if i == len(portList)-1 {
-			portList[i] = strings.Join(line, " ")
-		} else {
-			portList[i] = strings.Join(line, " ") + ` \`
-		}
-	}
-
-	dockerfile := `FROM scratch
-	EXPOSE {{range .}} {{.}}
-	{{end}}`
-	tmpl := template.Must(template.New("dockerfile").Parse(dockerfile))
-	buf := bytes.NewBuffer(nil)
-	tmpl.Execute(buf, portList)
-
-	name := "testbuildexpose"
-	_, err := buildImage(name, buf.String(), true)
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	// check if all the ports are saved inside Config.ExposedPorts
-	res, err := inspectFieldJSON(name, "Config.ExposedPorts")
-	if err != nil {
-		c.Fatal(err)
-	}
-	var exposedPorts map[string]interface{}
-	if err := json.Unmarshal([]byte(res), &exposedPorts); err != nil {
-		c.Fatal(err)
-	}
-
-	for _, p := range expectedPorts {
-		ep := fmt.Sprintf("%d/tcp", p)
-		if _, ok := exposedPorts[ep]; !ok {
-			c.Errorf("Port(%s) is not exposed", ep)
-		} else {
-			delete(exposedPorts, ep)
-		}
-	}
-	if len(exposedPorts) != 0 {
-		c.Errorf("Unexpected extra exposed ports %v", exposedPorts)
-	}
-}
-
-func (s *DockerSuite) TestBuildExposeOrder(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	buildID := func(name, exposed string) string {
-		_, err := buildImage(name, fmt.Sprintf(`FROM scratch
-		EXPOSE %s`, exposed), true)
-		if err != nil {
-			c.Fatal(err)
-		}
-		id, err := inspectField(name, "Id")
-		if err != nil {
-			c.Fatal(err)
-		}
-		return id
-	}
-
-	id1 := buildID("testbuildexpose1", "80 2375")
-	id2 := buildID("testbuildexpose2", "2375 80")
-	if id1 != id2 {
-		c.Errorf("EXPOSE should invalidate the cache only when ports actually changed")
-	}
-}
-
-func (s *DockerSuite) TestBuildExposeUpperCaseProto(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildexposeuppercaseproto"
-	expected := "map[5678/udp:{}]"
-	_, err := buildImage(name,
-		`FROM scratch
-        EXPOSE 5678/UDP`,
-		true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	res, err := inspectField(name, "Config.ExposedPorts")
-	if err != nil {
-		c.Fatal(err)
-	}
-	if res != expected {
-		c.Fatalf("Exposed ports %s, expected %s", res, expected)
 	}
 }
 
@@ -2522,76 +685,6 @@ func (s *DockerSuite) TestBuildEntrypoint(c *check.C) {
 	}
 	if res != expected {
 		c.Fatalf("Entrypoint %s, expected %s", res, expected)
-	}
-
-}
-
-// #6445 ensure ONBUILD triggers aren't committed to grandchildren
-func (s *DockerSuite) TestBuildOnBuildLimitedInheritence(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	var (
-		out2, out3 string
-	)
-	{
-		name1 := "testonbuildtrigger1"
-		dockerfile1 := `
-		FROM busybox
-		RUN echo "GRANDPARENT"
-		ONBUILD RUN echo "ONBUILD PARENT"
-		`
-		ctx, err := fakeContext(dockerfile1, nil)
-		if err != nil {
-			c.Fatal(err)
-		}
-		defer ctx.Close()
-
-		out1, _, err := dockerCmdInDir(c, ctx.Dir, "build", "-t", name1, ".")
-		if err != nil {
-			c.Fatalf("build failed to complete: %s, %v", out1, err)
-		}
-	}
-	{
-		name2 := "testonbuildtrigger2"
-		dockerfile2 := `
-		FROM testonbuildtrigger1
-		`
-		ctx, err := fakeContext(dockerfile2, nil)
-		if err != nil {
-			c.Fatal(err)
-		}
-		defer ctx.Close()
-
-		out2, _, err = dockerCmdInDir(c, ctx.Dir, "build", "-t", name2, ".")
-		if err != nil {
-			c.Fatalf("build failed to complete: %s, %v", out2, err)
-		}
-	}
-	{
-		name3 := "testonbuildtrigger3"
-		dockerfile3 := `
-		FROM testonbuildtrigger2
-		`
-		ctx, err := fakeContext(dockerfile3, nil)
-		if err != nil {
-			c.Fatal(err)
-		}
-		defer ctx.Close()
-
-		out3, _, err = dockerCmdInDir(c, ctx.Dir, "build", "-t", name3, ".")
-		if err != nil {
-			c.Fatalf("build failed to complete: %s, %v", out3, err)
-		}
-
-	}
-
-	// ONBUILD should be run in second build.
-	if !strings.Contains(out2, "ONBUILD PARENT") {
-		c.Fatalf("ONBUILD instruction did not run in child of ONBUILD parent")
-	}
-
-	// ONBUILD should *not* be run in third build.
-	if strings.Contains(out3, "ONBUILD PARENT") {
-		c.Fatalf("ONBUILD instruction ran in grandchild of ONBUILD parent")
 	}
 
 }
@@ -2691,462 +784,6 @@ func (s *DockerSuite) TestBuildConditionalCache(c *check.C) {
 	}
 }
 
-func (s *DockerSuite) TestBuildAddLocalFileWithCache(c *check.C) {
-	// local files are not owned by the correct user
-	testRequires(c, NotUserNamespace)
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildaddlocalfilewithcache"
-	name2 := "testbuildaddlocalfilewithcache2"
-	dockerfile := `
-		FROM busybox
-        MAINTAINER dockerio
-        ADD foo /usr/lib/bla/bar
-		RUN [ "$(cat /usr/lib/bla/bar)" = "hello" ]`
-	ctx, err := fakeContext(dockerfile, map[string]string{
-		"foo": "hello",
-	})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-	id1, err := buildImageFromContext(name, ctx, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	id2, err := buildImageFromContext(name2, ctx, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	if id1 != id2 {
-		c.Fatal("The cache should have been used but hasn't.")
-	}
-}
-
-func (s *DockerSuite) TestBuildAddMultipleLocalFileWithCache(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildaddmultiplelocalfilewithcache"
-	name2 := "testbuildaddmultiplelocalfilewithcache2"
-	dockerfile := `
-		FROM busybox
-        MAINTAINER dockerio
-        ADD foo Dockerfile /usr/lib/bla/
-		RUN [ "$(cat /usr/lib/bla/foo)" = "hello" ]`
-	ctx, err := fakeContext(dockerfile, map[string]string{
-		"foo": "hello",
-	})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-	id1, err := buildImageFromContext(name, ctx, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	id2, err := buildImageFromContext(name2, ctx, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	if id1 != id2 {
-		c.Fatal("The cache should have been used but hasn't.")
-	}
-}
-
-func (s *DockerSuite) TestBuildAddLocalFileWithoutCache(c *check.C) {
-	// local files are not owned by the correct user
-	testRequires(c, NotUserNamespace)
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildaddlocalfilewithoutcache"
-	name2 := "testbuildaddlocalfilewithoutcache2"
-	dockerfile := `
-		FROM busybox
-        MAINTAINER dockerio
-        ADD foo /usr/lib/bla/bar
-		RUN [ "$(cat /usr/lib/bla/bar)" = "hello" ]`
-	ctx, err := fakeContext(dockerfile, map[string]string{
-		"foo": "hello",
-	})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-	id1, err := buildImageFromContext(name, ctx, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	id2, err := buildImageFromContext(name2, ctx, false)
-	if err != nil {
-		c.Fatal(err)
-	}
-	if id1 == id2 {
-		c.Fatal("The cache should have been invalided but hasn't.")
-	}
-}
-
-func (s *DockerSuite) TestBuildCopyDirButNotFile(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildcopydirbutnotfile"
-	name2 := "testbuildcopydirbutnotfile2"
-	dockerfile := `
-        FROM scratch
-        COPY dir /tmp/`
-	ctx, err := fakeContext(dockerfile, map[string]string{
-		"dir/foo": "hello",
-	})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-	id1, err := buildImageFromContext(name, ctx, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	// Check that adding file with similar name doesn't mess with cache
-	if err := ctx.Add("dir_file", "hello2"); err != nil {
-		c.Fatal(err)
-	}
-	id2, err := buildImageFromContext(name2, ctx, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	if id1 != id2 {
-		c.Fatal("The cache should have been used but wasn't")
-	}
-}
-
-func (s *DockerSuite) TestBuildAddCurrentDirWithCache(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildaddcurrentdirwithcache"
-	name2 := name + "2"
-	name3 := name + "3"
-	name4 := name + "4"
-	dockerfile := `
-        FROM scratch
-        MAINTAINER dockerio
-        ADD . /usr/lib/bla`
-	ctx, err := fakeContext(dockerfile, map[string]string{
-		"foo": "hello",
-	})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-	id1, err := buildImageFromContext(name, ctx, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	// Check that adding file invalidate cache of "ADD ."
-	if err := ctx.Add("bar", "hello2"); err != nil {
-		c.Fatal(err)
-	}
-	id2, err := buildImageFromContext(name2, ctx, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	if id1 == id2 {
-		c.Fatal("The cache should have been invalided but hasn't.")
-	}
-	// Check that changing file invalidate cache of "ADD ."
-	if err := ctx.Add("foo", "hello1"); err != nil {
-		c.Fatal(err)
-	}
-	id3, err := buildImageFromContext(name3, ctx, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	if id2 == id3 {
-		c.Fatal("The cache should have been invalided but hasn't.")
-	}
-	// Check that changing file to same content with different mtime does not
-	// invalidate cache of "ADD ."
-	time.Sleep(1 * time.Second) // wait second because of mtime precision
-	if err := ctx.Add("foo", "hello1"); err != nil {
-		c.Fatal(err)
-	}
-	id4, err := buildImageFromContext(name4, ctx, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	if id3 != id4 {
-		c.Fatal("The cache should have been used but hasn't.")
-	}
-}
-
-func (s *DockerSuite) TestBuildAddCurrentDirWithoutCache(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildaddcurrentdirwithoutcache"
-	name2 := "testbuildaddcurrentdirwithoutcache2"
-	dockerfile := `
-        FROM scratch
-        MAINTAINER dockerio
-        ADD . /usr/lib/bla`
-	ctx, err := fakeContext(dockerfile, map[string]string{
-		"foo": "hello",
-	})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-	id1, err := buildImageFromContext(name, ctx, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	id2, err := buildImageFromContext(name2, ctx, false)
-	if err != nil {
-		c.Fatal(err)
-	}
-	if id1 == id2 {
-		c.Fatal("The cache should have been invalided but hasn't.")
-	}
-}
-
-func (s *DockerSuite) TestBuildAddRemoteFileWithCache(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildaddremotefilewithcache"
-	server, err := fakeStorage(map[string]string{
-		"baz": "hello",
-	})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer server.Close()
-
-	id1, err := buildImage(name,
-		fmt.Sprintf(`FROM scratch
-        MAINTAINER dockerio
-        ADD %s/baz /usr/lib/baz/quux`, server.URL()),
-		true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	id2, err := buildImage(name,
-		fmt.Sprintf(`FROM scratch
-        MAINTAINER dockerio
-        ADD %s/baz /usr/lib/baz/quux`, server.URL()),
-		true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	if id1 != id2 {
-		c.Fatal("The cache should have been used but hasn't.")
-	}
-}
-
-func (s *DockerSuite) TestBuildAddRemoteFileWithoutCache(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildaddremotefilewithoutcache"
-	name2 := "testbuildaddremotefilewithoutcache2"
-	server, err := fakeStorage(map[string]string{
-		"baz": "hello",
-	})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer server.Close()
-
-	id1, err := buildImage(name,
-		fmt.Sprintf(`FROM scratch
-        MAINTAINER dockerio
-        ADD %s/baz /usr/lib/baz/quux`, server.URL()),
-		true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	id2, err := buildImage(name2,
-		fmt.Sprintf(`FROM scratch
-        MAINTAINER dockerio
-        ADD %s/baz /usr/lib/baz/quux`, server.URL()),
-		false)
-	if err != nil {
-		c.Fatal(err)
-	}
-	if id1 == id2 {
-		c.Fatal("The cache should have been invalided but hasn't.")
-	}
-}
-
-func (s *DockerSuite) TestBuildAddRemoteFileMTime(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildaddremotefilemtime"
-	name2 := name + "2"
-	name3 := name + "3"
-
-	files := map[string]string{"baz": "hello"}
-	server, err := fakeStorage(files)
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer server.Close()
-
-	ctx, err := fakeContext(fmt.Sprintf(`FROM scratch
-        MAINTAINER dockerio
-        ADD %s/baz /usr/lib/baz/quux`, server.URL()), nil)
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	id1, err := buildImageFromContext(name, ctx, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	id2, err := buildImageFromContext(name2, ctx, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	if id1 != id2 {
-		c.Fatal("The cache should have been used but wasn't - #1")
-	}
-
-	// Now create a different server with same contents (causes different mtime)
-	// The cache should still be used
-
-	// allow some time for clock to pass as mtime precision is only 1s
-	time.Sleep(2 * time.Second)
-
-	server2, err := fakeStorage(files)
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer server2.Close()
-
-	ctx2, err := fakeContext(fmt.Sprintf(`FROM scratch
-        MAINTAINER dockerio
-        ADD %s/baz /usr/lib/baz/quux`, server2.URL()), nil)
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx2.Close()
-	id3, err := buildImageFromContext(name3, ctx2, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	if id1 != id3 {
-		c.Fatal("The cache should have been used but wasn't")
-	}
-}
-
-func (s *DockerSuite) TestBuildAddLocalAndRemoteFilesWithCache(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildaddlocalandremotefilewithcache"
-	server, err := fakeStorage(map[string]string{
-		"baz": "hello",
-	})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer server.Close()
-
-	ctx, err := fakeContext(fmt.Sprintf(`FROM scratch
-        MAINTAINER dockerio
-        ADD foo /usr/lib/bla/bar
-        ADD %s/baz /usr/lib/baz/quux`, server.URL()),
-		map[string]string{
-			"foo": "hello world",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-	id1, err := buildImageFromContext(name, ctx, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	id2, err := buildImageFromContext(name, ctx, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	if id1 != id2 {
-		c.Fatal("The cache should have been used but hasn't.")
-	}
-}
-
-func testContextTar(c *check.C, compression archive.Compression) {
-	testRequires(c, DaemonIsLinux)
-	ctx, err := fakeContext(
-		`FROM busybox
-ADD foo /foo
-CMD ["cat", "/foo"]`,
-		map[string]string{
-			"foo": "bar",
-		},
-	)
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-	context, err := archive.Tar(ctx.Dir, compression)
-	if err != nil {
-		c.Fatalf("failed to build context tar: %v", err)
-	}
-	name := "contexttar"
-	buildCmd := exec.Command(dockerBinary, "build", "-t", name, "-")
-	buildCmd.Stdin = context
-
-	if out, _, err := runCommandWithOutput(buildCmd); err != nil {
-		c.Fatalf("build failed to complete: %v %v", out, err)
-	}
-}
-
-func (s *DockerSuite) TestBuildContextTarGzip(c *check.C) {
-	testContextTar(c, archive.Gzip)
-}
-
-func (s *DockerSuite) TestBuildContextTarNoCompression(c *check.C) {
-	testContextTar(c, archive.Uncompressed)
-}
-
-func (s *DockerSuite) TestBuildNoContext(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	buildCmd := exec.Command(dockerBinary, "build", "-t", "nocontext", "-")
-	buildCmd.Stdin = strings.NewReader("FROM busybox\nCMD echo ok\n")
-
-	if out, _, err := runCommandWithOutput(buildCmd); err != nil {
-		c.Fatalf("build failed to complete: %v %v", out, err)
-	}
-
-	if out, _ := dockerCmd(c, "run", "--rm", "nocontext"); out != "ok\n" {
-		c.Fatalf("run produced invalid output: %q, expected %q", out, "ok")
-	}
-}
-
-// TODO: TestCaching
-func (s *DockerSuite) TestBuildAddLocalAndRemoteFilesWithoutCache(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildaddlocalandremotefilewithoutcache"
-	name2 := "testbuildaddlocalandremotefilewithoutcache2"
-	server, err := fakeStorage(map[string]string{
-		"baz": "hello",
-	})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer server.Close()
-
-	ctx, err := fakeContext(fmt.Sprintf(`FROM scratch
-        MAINTAINER dockerio
-        ADD foo /usr/lib/bla/bar
-        ADD %s/baz /usr/lib/baz/quux`, server.URL()),
-		map[string]string{
-			"foo": "hello world",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-	id1, err := buildImageFromContext(name, ctx, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	id2, err := buildImageFromContext(name2, ctx, false)
-	if err != nil {
-		c.Fatal(err)
-	}
-	if id1 == id2 {
-		c.Fatal("The cache should have been invalided but hasn't.")
-	}
-}
-
 func (s *DockerSuite) TestBuildWithVolumeOwnership(c *check.C) {
 	testRequires(c, DaemonIsLinux)
 	name := "testbuildimg"
@@ -3206,47 +843,6 @@ func (s *DockerSuite) TestBuildEntrypointRunCleanup(c *check.C) {
 	// Cmd must be cleaned up
 	if res != "<nil>" {
 		c.Fatalf("Cmd %s, expected nil", res)
-	}
-}
-
-func (s *DockerSuite) TestBuildForbiddenContextPath(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildforbidpath"
-	ctx, err := fakeContext(`FROM scratch
-        ADD ../../ test/
-        `,
-		map[string]string{
-			"test.txt":  "test1",
-			"other.txt": "other",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	expected := "Forbidden path outside the build context: ../../ "
-	if _, err := buildImageFromContext(name, ctx, true); err == nil || !strings.Contains(err.Error(), expected) {
-		c.Fatalf("Wrong error: (should contain \"%s\") got:\n%v", expected, err)
-	}
-
-}
-
-func (s *DockerSuite) TestBuildAddFileNotFound(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildaddnotfound"
-	ctx, err := fakeContext(`FROM scratch
-        ADD foo /usr/local/bar`,
-		map[string]string{"bar": "hello"})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		if !strings.Contains(err.Error(), "foo: no such file or directory") {
-			c.Fatalf("Wrong error %v, must be about missing foo file or directory", err)
-		}
-	} else {
-		c.Fatal("Error must not be nil")
 	}
 }
 
@@ -3318,95 +914,6 @@ func (s *DockerSuite) TestBuildFailsDockerfileEmpty(c *check.C) {
 	}
 }
 
-func (s *DockerSuite) TestBuildOnBuild(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildonbuild"
-	_, err := buildImage(name,
-		`FROM busybox
-		ONBUILD RUN touch foobar`,
-		true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	_, err = buildImage(name,
-		fmt.Sprintf(`FROM %s
-		RUN [ -f foobar ]`, name),
-		true)
-	if err != nil {
-		c.Fatal(err)
-	}
-}
-
-func (s *DockerSuite) TestBuildOnBuildForbiddenChained(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildonbuildforbiddenchained"
-	_, err := buildImage(name,
-		`FROM busybox
-		ONBUILD ONBUILD RUN touch foobar`,
-		true)
-	if err != nil {
-		if !strings.Contains(err.Error(), "Chaining ONBUILD via `ONBUILD ONBUILD` isn't allowed") {
-			c.Fatalf("Wrong error %v, must be about chaining ONBUILD", err)
-		}
-	} else {
-		c.Fatal("Error must not be nil")
-	}
-}
-
-func (s *DockerSuite) TestBuildOnBuildForbiddenFrom(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildonbuildforbiddenfrom"
-	_, err := buildImage(name,
-		`FROM busybox
-		ONBUILD FROM scratch`,
-		true)
-	if err != nil {
-		if !strings.Contains(err.Error(), "FROM isn't allowed as an ONBUILD trigger") {
-			c.Fatalf("Wrong error %v, must be about FROM forbidden", err)
-		}
-	} else {
-		c.Fatal("Error must not be nil")
-	}
-}
-
-func (s *DockerSuite) TestBuildOnBuildForbiddenMaintainer(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildonbuildforbiddenmaintainer"
-	_, err := buildImage(name,
-		`FROM busybox
-		ONBUILD MAINTAINER docker.io`,
-		true)
-	if err != nil {
-		if !strings.Contains(err.Error(), "MAINTAINER isn't allowed as an ONBUILD trigger") {
-			c.Fatalf("Wrong error %v, must be about MAINTAINER forbidden", err)
-		}
-	} else {
-		c.Fatal("Error must not be nil")
-	}
-}
-
-// gh #2446
-func (s *DockerSuite) TestBuildAddToSymlinkDest(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildaddtosymlinkdest"
-	ctx, err := fakeContext(`FROM busybox
-        RUN mkdir /foo
-        RUN ln -s /foo /bar
-        ADD foo /bar/
-        RUN [ -f /bar/foo ]
-        RUN [ -f /foo/foo ]`,
-		map[string]string{
-			"foo": "hello",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-}
-
 func (s *DockerSuite) TestBuildEscapeWhitespace(c *check.C) {
 	testRequires(c, DaemonIsLinux)
 	name := "testbuildescaping"
@@ -3453,428 +960,6 @@ func (s *DockerSuite) TestBuildVerifyIntString(c *check.C) {
 		c.Fatalf("Output does not contain the int as a string:\n%s", out)
 	}
 
-}
-
-func (s *DockerSuite) TestBuildDockerignore(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuilddockerignore"
-	dockerfile := `
-        FROM busybox
-        ADD . /bla
-		RUN [[ -f /bla/src/x.go ]]
-		RUN [[ -f /bla/Makefile ]]
-		RUN [[ ! -e /bla/src/_vendor ]]
-		RUN [[ ! -e /bla/.gitignore ]]
-		RUN [[ ! -e /bla/README.md ]]
-		RUN [[ ! -e /bla/dir/foo ]]
-		RUN [[ ! -e /bla/foo ]]
-		RUN [[ ! -e /bla/.git ]]
-		RUN [[ ! -e v.cc ]]
-		RUN [[ ! -e src/v.cc ]]
-		RUN [[ ! -e src/_vendor/v.cc ]]`
-	ctx, err := fakeContext(dockerfile, map[string]string{
-		"Makefile":         "all:",
-		".git/HEAD":        "ref: foo",
-		"src/x.go":         "package main",
-		"src/_vendor/v.go": "package main",
-		"src/_vendor/v.cc": "package main",
-		"src/v.cc":         "package main",
-		"v.cc":             "package main",
-		"dir/foo":          "",
-		".gitignore":       "",
-		"README.md":        "readme",
-		".dockerignore": `
-.git
-pkg
-.gitignore
-src/_vendor
-*.md
-**/*.cc
-dir`,
-	})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-}
-
-func (s *DockerSuite) TestBuildDockerignoreCleanPaths(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuilddockerignorecleanpaths"
-	dockerfile := `
-        FROM busybox
-        ADD . /tmp/
-        RUN (! ls /tmp/foo) && (! ls /tmp/foo2) && (! ls /tmp/dir1/foo)`
-	ctx, err := fakeContext(dockerfile, map[string]string{
-		"foo":           "foo",
-		"foo2":          "foo2",
-		"dir1/foo":      "foo in dir1",
-		".dockerignore": "./foo\ndir1//foo\n./dir1/../foo2",
-	})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-}
-
-func (s *DockerSuite) TestBuildDockerignoreExceptions(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuilddockerignoreexceptions"
-	dockerfile := `
-        FROM busybox
-        ADD . /bla
-		RUN [[ -f /bla/src/x.go ]]
-		RUN [[ -f /bla/Makefile ]]
-		RUN [[ ! -e /bla/src/_vendor ]]
-		RUN [[ ! -e /bla/.gitignore ]]
-		RUN [[ ! -e /bla/README.md ]]
-		RUN [[  -e /bla/dir/dir/foo ]]
-		RUN [[ ! -e /bla/dir/foo1 ]]
-		RUN [[ -f /bla/dir/e ]]
-		RUN [[ -f /bla/dir/e-dir/foo ]]
-		RUN [[ ! -e /bla/foo ]]
-		RUN [[ ! -e /bla/.git ]]
-		RUN [[ -e /bla/dir/a.cc ]]`
-	ctx, err := fakeContext(dockerfile, map[string]string{
-		"Makefile":         "all:",
-		".git/HEAD":        "ref: foo",
-		"src/x.go":         "package main",
-		"src/_vendor/v.go": "package main",
-		"dir/foo":          "",
-		"dir/foo1":         "",
-		"dir/dir/f1":       "",
-		"dir/dir/foo":      "",
-		"dir/e":            "",
-		"dir/e-dir/foo":    "",
-		".gitignore":       "",
-		"README.md":        "readme",
-		"dir/a.cc":         "hello",
-		".dockerignore": `
-.git
-pkg
-.gitignore
-src/_vendor
-*.md
-dir
-!dir/e*
-!dir/dir/foo
-**/*.cc
-!**/*.cc`,
-	})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-}
-
-func (s *DockerSuite) TestBuildDockerignoringDockerfile(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuilddockerignoredockerfile"
-	dockerfile := `
-        FROM busybox
-		ADD . /tmp/
-		RUN ! ls /tmp/Dockerfile
-		RUN ls /tmp/.dockerignore`
-	ctx, err := fakeContext(dockerfile, map[string]string{
-		"Dockerfile":    dockerfile,
-		".dockerignore": "Dockerfile\n",
-	})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	if _, err = buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatalf("Didn't ignore Dockerfile correctly:%s", err)
-	}
-
-	// now try it with ./Dockerfile
-	ctx.Add(".dockerignore", "./Dockerfile\n")
-	if _, err = buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatalf("Didn't ignore ./Dockerfile correctly:%s", err)
-	}
-
-}
-
-func (s *DockerSuite) TestBuildDockerignoringRenamedDockerfile(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuilddockerignoredockerfile"
-	dockerfile := `
-        FROM busybox
-		ADD . /tmp/
-		RUN ls /tmp/Dockerfile
-		RUN ! ls /tmp/MyDockerfile
-		RUN ls /tmp/.dockerignore`
-	ctx, err := fakeContext(dockerfile, map[string]string{
-		"Dockerfile":    "Should not use me",
-		"MyDockerfile":  dockerfile,
-		".dockerignore": "MyDockerfile\n",
-	})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	if _, err = buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatalf("Didn't ignore MyDockerfile correctly:%s", err)
-	}
-
-	// now try it with ./MyDockerfile
-	ctx.Add(".dockerignore", "./MyDockerfile\n")
-	if _, err = buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatalf("Didn't ignore ./MyDockerfile correctly:%s", err)
-	}
-
-}
-
-func (s *DockerSuite) TestBuildDockerignoringDockerignore(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuilddockerignoredockerignore"
-	dockerfile := `
-        FROM busybox
-		ADD . /tmp/
-		RUN ! ls /tmp/.dockerignore
-		RUN ls /tmp/Dockerfile`
-	ctx, err := fakeContext(dockerfile, map[string]string{
-		"Dockerfile":    dockerfile,
-		".dockerignore": ".dockerignore\n",
-	})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-	if _, err = buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatalf("Didn't ignore .dockerignore correctly:%s", err)
-	}
-}
-
-func (s *DockerSuite) TestBuildDockerignoreTouchDockerfile(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	var id1 string
-	var id2 string
-
-	name := "testbuilddockerignoretouchdockerfile"
-	dockerfile := `
-        FROM busybox
-		ADD . /tmp/`
-	ctx, err := fakeContext(dockerfile, map[string]string{
-		"Dockerfile":    dockerfile,
-		".dockerignore": "Dockerfile\n",
-	})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	if id1, err = buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatalf("Didn't build it correctly:%s", err)
-	}
-
-	if id2, err = buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatalf("Didn't build it correctly:%s", err)
-	}
-	if id1 != id2 {
-		c.Fatalf("Didn't use the cache - 1")
-	}
-
-	// Now make sure touching Dockerfile doesn't invalidate the cache
-	if err = ctx.Add("Dockerfile", dockerfile+"\n# hi"); err != nil {
-		c.Fatalf("Didn't add Dockerfile: %s", err)
-	}
-	if id2, err = buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatalf("Didn't build it correctly:%s", err)
-	}
-	if id1 != id2 {
-		c.Fatalf("Didn't use the cache - 2")
-	}
-
-	// One more time but just 'touch' it instead of changing the content
-	if err = ctx.Add("Dockerfile", dockerfile+"\n# hi"); err != nil {
-		c.Fatalf("Didn't add Dockerfile: %s", err)
-	}
-	if id2, err = buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatalf("Didn't build it correctly:%s", err)
-	}
-	if id1 != id2 {
-		c.Fatalf("Didn't use the cache - 3")
-	}
-
-}
-
-func (s *DockerSuite) TestBuildDockerignoringWholeDir(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuilddockerignorewholedir"
-	dockerfile := `
-        FROM busybox
-		COPY . /
-		RUN [[ ! -e /.gitignore ]]
-		RUN [[ -f /Makefile ]]`
-	ctx, err := fakeContext(dockerfile, map[string]string{
-		"Dockerfile":    "FROM scratch",
-		"Makefile":      "all:",
-		".gitignore":    "",
-		".dockerignore": ".*\n",
-	})
-	c.Assert(err, check.IsNil)
-	defer ctx.Close()
-	if _, err = buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-
-	c.Assert(ctx.Add(".dockerfile", "*"), check.IsNil)
-	if _, err = buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-
-	c.Assert(ctx.Add(".dockerfile", "."), check.IsNil)
-	if _, err = buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-
-	c.Assert(ctx.Add(".dockerfile", "?"), check.IsNil)
-	if _, err = buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatal(err)
-	}
-}
-
-func (s *DockerSuite) TestBuildDockerignoringBadExclusion(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuilddockerignorebadexclusion"
-	dockerfile := `
-        FROM busybox
-		COPY . /
-		RUN [[ ! -e /.gitignore ]]
-		RUN [[ -f /Makefile ]]`
-	ctx, err := fakeContext(dockerfile, map[string]string{
-		"Dockerfile":    "FROM scratch",
-		"Makefile":      "all:",
-		".gitignore":    "",
-		".dockerignore": "!\n",
-	})
-	c.Assert(err, check.IsNil)
-	defer ctx.Close()
-	if _, err = buildImageFromContext(name, ctx, true); err == nil {
-		c.Fatalf("Build was supposed to fail but didn't")
-	}
-
-	if err.Error() != "failed to build the image: Error checking context: 'Illegal exclusion pattern: !'.\n" {
-		c.Fatalf("Incorrect output, got:%q", err.Error())
-	}
-}
-
-func (s *DockerSuite) TestBuildDockerignoringWildTopDir(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-
-	dockerfile := `
-        FROM busybox
-		COPY . /
-		RUN [[ ! -e /.dockerignore ]]
-		RUN [[ ! -e /Dockerfile ]]
-		RUN [[ ! -e /file1 ]]
-		RUN [[ ! -e /dir ]]`
-
-	ctx, err := fakeContext(dockerfile, map[string]string{
-		"Dockerfile": "FROM scratch",
-		"file1":      "",
-		"dir/dfile1": "",
-	})
-	c.Assert(err, check.IsNil)
-	defer ctx.Close()
-
-	// All of these should result in ignoring all files
-	for _, variant := range []string{"**", "**/", "**/**", "*"} {
-		ctx.Add(".dockerignore", variant)
-		_, err = buildImageFromContext("noname", ctx, true)
-		c.Assert(err, check.IsNil, check.Commentf("variant: %s", variant))
-	}
-}
-
-func (s *DockerSuite) TestBuildDockerignoringWildDirs(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-
-	dockerfile := `
-        FROM busybox
-		COPY . /
-		RUN [[ -e /.dockerignore ]]
-		RUN [[ -e /Dockerfile ]]
-
-		RUN [[ ! -e /file0 ]]
-		RUN [[ ! -e /dir1/file0 ]]
-		RUN [[ ! -e /dir2/file0 ]]
-
-		RUN [[ ! -e /file1 ]]
-		RUN [[ ! -e /dir1/file1 ]]
-		RUN [[ ! -e /dir1/dir2/file1 ]]
-
-		RUN [[ ! -e /dir1/file2 ]]
-		RUN [[   -e /dir1/dir2/file2 ]]
-
-		RUN [[ ! -e /dir1/dir2/file4 ]]
-		RUN [[ ! -e /dir1/dir2/file5 ]]
-		RUN [[ ! -e /dir1/dir2/file6 ]]
-		RUN [[ ! -e /dir1/dir3/file7 ]]
-		RUN [[ ! -e /dir1/dir3/file8 ]]
-		RUN [[   -e /dir1/dir3 ]]
-		RUN [[   -e /dir1/dir4 ]]
-
-		RUN [[ ! -e 'dir1/dir5/fileAA' ]]
-		RUN [[   -e 'dir1/dir5/fileAB' ]]
-		RUN [[   -e 'dir1/dir5/fileB' ]]   # "." in pattern means nothing
-
-		RUN echo all done!`
-
-	ctx, err := fakeContext(dockerfile, map[string]string{
-		"Dockerfile":      "FROM scratch",
-		"file0":           "",
-		"dir1/file0":      "",
-		"dir1/dir2/file0": "",
-
-		"file1":           "",
-		"dir1/file1":      "",
-		"dir1/dir2/file1": "",
-
-		"dir1/file2":      "",
-		"dir1/dir2/file2": "", // remains
-
-		"dir1/dir2/file4": "",
-		"dir1/dir2/file5": "",
-		"dir1/dir2/file6": "",
-		"dir1/dir3/file7": "",
-		"dir1/dir3/file8": "",
-		"dir1/dir4/file9": "",
-
-		"dir1/dir5/fileAA": "",
-		"dir1/dir5/fileAB": "",
-		"dir1/dir5/fileB":  "",
-
-		".dockerignore": `
-**/file0
-**/*file1
-**/dir1/file2
-dir1/**/file4
-**/dir2/file5
-**/dir1/dir2/file6
-dir1/dir3/**
-**/dir4/**
-**/file?A
-**/file\?B
-**/dir5/file.
-`,
-	})
-	c.Assert(err, check.IsNil)
-	defer ctx.Close()
-
-	_, err = buildImageFromContext("noname", ctx, true)
-	c.Assert(err, check.IsNil)
 }
 
 func (s *DockerSuite) TestBuildLineBreak(c *check.C) {
@@ -3991,538 +1076,6 @@ RUN [ "$(id -u):$(id -g)/$(id -un):$(id -gn)/$(id -G):$(id -Gn)" = '1042:1043/10
 	}
 }
 
-func (s *DockerSuite) TestBuildEnvUsage(c *check.C) {
-	// /docker/world/hello is not owned by the correct user
-	testRequires(c, NotUserNamespace)
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildenvusage"
-	dockerfile := `FROM busybox
-ENV    HOME /root
-ENV    PATH $HOME/bin:$PATH
-ENV    PATH /tmp:$PATH
-RUN    [ "$PATH" = "/tmp:$HOME/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin" ]
-ENV    FOO /foo/baz
-ENV    BAR /bar
-ENV    BAZ $BAR
-ENV    FOOPATH $PATH:$FOO
-RUN    [ "$BAR" = "$BAZ" ]
-RUN    [ "$FOOPATH" = "$PATH:/foo/baz" ]
-ENV	   FROM hello/docker/world
-ENV    TO /docker/world/hello
-ADD    $FROM $TO
-RUN    [ "$(cat $TO)" = "hello" ]
-ENV    abc=def
-ENV    ghi=$abc
-RUN    [ "$ghi" = "def" ]
-`
-	ctx, err := fakeContext(dockerfile, map[string]string{
-		"hello/docker/world": "hello",
-	})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	_, err = buildImageFromContext(name, ctx, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-}
-
-func (s *DockerSuite) TestBuildEnvUsage2(c *check.C) {
-	// /docker/world/hello is not owned by the correct user
-	testRequires(c, NotUserNamespace)
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildenvusage2"
-	dockerfile := `FROM busybox
-ENV    abc=def
-RUN    [ "$abc" = "def" ]
-ENV    def="hello world"
-RUN    [ "$def" = "hello world" ]
-ENV    def=hello\ world
-RUN    [ "$def" = "hello world" ]
-ENV    v1=abc v2="hi there"
-RUN    [ "$v1" = "abc" ]
-RUN    [ "$v2" = "hi there" ]
-ENV    v3='boogie nights' v4="with'quotes too"
-RUN    [ "$v3" = "boogie nights" ]
-RUN    [ "$v4" = "with'quotes too" ]
-ENV    abc=zzz FROM=hello/docker/world
-ENV    abc=zzz TO=/docker/world/hello
-ADD    $FROM $TO
-RUN    [ "$(cat $TO)" = "hello" ]
-ENV    abc "zzz"
-RUN    [ $abc = "zzz" ]
-ENV    abc 'yyy'
-RUN    [ $abc = 'yyy' ]
-ENV    abc=
-RUN    [ "$abc" = "" ]
-
-# use grep to make sure if the builder substitutes \$foo by mistake
-# we don't get a false positive
-ENV    abc=\$foo
-RUN    [ "$abc" = "\$foo" ] && (echo "$abc" | grep foo)
-ENV    abc \$foo
-RUN    [ "$abc" = "\$foo" ] && (echo "$abc" | grep foo)
-
-ENV    abc=\'foo\'
-RUN    [ "$abc" = "'foo'" ]
-ENV    abc=\"foo\"
-RUN    [ "$abc" = "\"foo\"" ]
-ENV    abc "foo"
-RUN    [ "$abc" = "foo" ]
-ENV    abc 'foo'
-RUN    [ "$abc" = 'foo' ]
-ENV    abc \'foo\'
-RUN    [ "$abc" = "'foo'" ]
-ENV    abc \"foo\"
-RUN    [ "$abc" = '"foo"' ]
-
-ENV    abc=ABC
-RUN    [ "$abc" = "ABC" ]
-ENV    def=${abc:-DEF}
-RUN    [ "$def" = "ABC" ]
-ENV    def=${ccc:-DEF}
-RUN    [ "$def" = "DEF" ]
-ENV    def=${ccc:-${def}xx}
-RUN    [ "$def" = "DEFxx" ]
-ENV    def=${def:+ALT}
-RUN    [ "$def" = "ALT" ]
-ENV    def=${def:+${abc}:}
-RUN    [ "$def" = "ABC:" ]
-ENV    def=${ccc:-\$abc:}
-RUN    [ "$def" = '$abc:' ]
-ENV    def=${ccc:-\${abc}:}
-RUN    [ "$def" = '${abc:}' ]
-ENV    mypath=${mypath:+$mypath:}/home
-RUN    [ "$mypath" = '/home' ]
-ENV    mypath=${mypath:+$mypath:}/away
-RUN    [ "$mypath" = '/home:/away' ]
-
-ENV    e1=bar
-ENV    e2=$e1
-ENV    e3=$e11
-ENV    e4=\$e1
-ENV    e5=\$e11
-RUN    [ "$e0,$e1,$e2,$e3,$e4,$e5" = ',bar,bar,,$e1,$e11' ]
-
-ENV    ee1 bar
-ENV    ee2 $ee1
-ENV    ee3 $ee11
-ENV    ee4 \$ee1
-ENV    ee5 \$ee11
-RUN    [ "$ee1,$ee2,$ee3,$ee4,$ee5" = 'bar,bar,,$ee1,$ee11' ]
-
-ENV    eee1="foo"
-ENV    eee2='foo'
-ENV    eee3 "foo"
-ENV    eee4 'foo'
-RUN    [ "$eee1,$eee2,$eee3,$eee4" = 'foo,foo,foo,foo' ]
-
-`
-	ctx, err := fakeContext(dockerfile, map[string]string{
-		"hello/docker/world": "hello",
-	})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	_, err = buildImageFromContext(name, ctx, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-}
-
-func (s *DockerSuite) TestBuildAddScript(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildaddscript"
-	dockerfile := `
-FROM busybox
-ADD test /test
-RUN ["chmod","+x","/test"]
-RUN ["/test"]
-RUN [ "$(cat /testfile)" = 'test!' ]`
-	ctx, err := fakeContext(dockerfile, map[string]string{
-		"test": "#!/bin/sh\necho 'test!' > /testfile",
-	})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	_, err = buildImageFromContext(name, ctx, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-}
-
-func (s *DockerSuite) TestBuildAddTar(c *check.C) {
-	// /test/foo is not owned by the correct user
-	testRequires(c, NotUserNamespace)
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildaddtar"
-
-	ctx := func() *FakeContext {
-		dockerfile := `
-FROM busybox
-ADD test.tar /
-RUN cat /test/foo | grep Hi
-ADD test.tar /test.tar
-RUN cat /test.tar/test/foo | grep Hi
-ADD test.tar /unlikely-to-exist
-RUN cat /unlikely-to-exist/test/foo | grep Hi
-ADD test.tar /unlikely-to-exist-trailing-slash/
-RUN cat /unlikely-to-exist-trailing-slash/test/foo | grep Hi
-RUN mkdir /existing-directory
-ADD test.tar /existing-directory
-RUN cat /existing-directory/test/foo | grep Hi
-ADD test.tar /existing-directory-trailing-slash/
-RUN cat /existing-directory-trailing-slash/test/foo | grep Hi`
-		tmpDir, err := ioutil.TempDir("", "fake-context")
-		c.Assert(err, check.IsNil)
-		testTar, err := os.Create(filepath.Join(tmpDir, "test.tar"))
-		if err != nil {
-			c.Fatalf("failed to create test.tar archive: %v", err)
-		}
-		defer testTar.Close()
-
-		tw := tar.NewWriter(testTar)
-
-		if err := tw.WriteHeader(&tar.Header{
-			Name: "test/foo",
-			Size: 2,
-		}); err != nil {
-			c.Fatalf("failed to write tar file header: %v", err)
-		}
-		if _, err := tw.Write([]byte("Hi")); err != nil {
-			c.Fatalf("failed to write tar file content: %v", err)
-		}
-		if err := tw.Close(); err != nil {
-			c.Fatalf("failed to close tar archive: %v", err)
-		}
-
-		if err := ioutil.WriteFile(filepath.Join(tmpDir, "Dockerfile"), []byte(dockerfile), 0644); err != nil {
-			c.Fatalf("failed to open destination dockerfile: %v", err)
-		}
-		return fakeContextFromDir(tmpDir)
-	}()
-	defer ctx.Close()
-
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatalf("build failed to complete for TestBuildAddTar: %v", err)
-	}
-
-}
-
-func (s *DockerSuite) TestBuildAddBrokenTar(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildaddbrokentar"
-
-	ctx := func() *FakeContext {
-		dockerfile := `
-FROM busybox
-ADD test.tar /`
-		tmpDir, err := ioutil.TempDir("", "fake-context")
-		c.Assert(err, check.IsNil)
-		testTar, err := os.Create(filepath.Join(tmpDir, "test.tar"))
-		if err != nil {
-			c.Fatalf("failed to create test.tar archive: %v", err)
-		}
-		defer testTar.Close()
-
-		tw := tar.NewWriter(testTar)
-
-		if err := tw.WriteHeader(&tar.Header{
-			Name: "test/foo",
-			Size: 2,
-		}); err != nil {
-			c.Fatalf("failed to write tar file header: %v", err)
-		}
-		if _, err := tw.Write([]byte("Hi")); err != nil {
-			c.Fatalf("failed to write tar file content: %v", err)
-		}
-		if err := tw.Close(); err != nil {
-			c.Fatalf("failed to close tar archive: %v", err)
-		}
-
-		// Corrupt the tar by removing one byte off the end
-		stat, err := testTar.Stat()
-		if err != nil {
-			c.Fatalf("failed to stat tar archive: %v", err)
-		}
-		if err := testTar.Truncate(stat.Size() - 1); err != nil {
-			c.Fatalf("failed to truncate tar archive: %v", err)
-		}
-
-		if err := ioutil.WriteFile(filepath.Join(tmpDir, "Dockerfile"), []byte(dockerfile), 0644); err != nil {
-			c.Fatalf("failed to open destination dockerfile: %v", err)
-		}
-		return fakeContextFromDir(tmpDir)
-	}()
-	defer ctx.Close()
-
-	if _, err := buildImageFromContext(name, ctx, true); err == nil {
-		c.Fatalf("build should have failed for TestBuildAddBrokenTar")
-	}
-}
-
-func (s *DockerSuite) TestBuildAddNonTar(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildaddnontar"
-
-	// Should not try to extract test.tar
-	ctx, err := fakeContext(`
-		FROM busybox
-		ADD test.tar /
-		RUN test -f /test.tar`,
-		map[string]string{"test.tar": "not_a_tar_file"})
-
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatalf("build failed for TestBuildAddNonTar")
-	}
-}
-
-func (s *DockerSuite) TestBuildAddTarXz(c *check.C) {
-	// /test/foo is not owned by the correct user
-	testRequires(c, NotUserNamespace)
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildaddtarxz"
-
-	ctx := func() *FakeContext {
-		dockerfile := `
-			FROM busybox
-			ADD test.tar.xz /
-			RUN cat /test/foo | grep Hi`
-		tmpDir, err := ioutil.TempDir("", "fake-context")
-		c.Assert(err, check.IsNil)
-		testTar, err := os.Create(filepath.Join(tmpDir, "test.tar"))
-		if err != nil {
-			c.Fatalf("failed to create test.tar archive: %v", err)
-		}
-		defer testTar.Close()
-
-		tw := tar.NewWriter(testTar)
-
-		if err := tw.WriteHeader(&tar.Header{
-			Name: "test/foo",
-			Size: 2,
-		}); err != nil {
-			c.Fatalf("failed to write tar file header: %v", err)
-		}
-		if _, err := tw.Write([]byte("Hi")); err != nil {
-			c.Fatalf("failed to write tar file content: %v", err)
-		}
-		if err := tw.Close(); err != nil {
-			c.Fatalf("failed to close tar archive: %v", err)
-		}
-
-		xzCompressCmd := exec.Command("xz", "-k", "test.tar")
-		xzCompressCmd.Dir = tmpDir
-		out, _, err := runCommandWithOutput(xzCompressCmd)
-		if err != nil {
-			c.Fatal(err, out)
-		}
-
-		if err := ioutil.WriteFile(filepath.Join(tmpDir, "Dockerfile"), []byte(dockerfile), 0644); err != nil {
-			c.Fatalf("failed to open destination dockerfile: %v", err)
-		}
-		return fakeContextFromDir(tmpDir)
-	}()
-
-	defer ctx.Close()
-
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatalf("build failed to complete for TestBuildAddTarXz: %v", err)
-	}
-
-}
-
-func (s *DockerSuite) TestBuildAddTarXzGz(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildaddtarxzgz"
-
-	ctx := func() *FakeContext {
-		dockerfile := `
-			FROM busybox
-			ADD test.tar.xz.gz /
-			RUN ls /test.tar.xz.gz`
-		tmpDir, err := ioutil.TempDir("", "fake-context")
-		c.Assert(err, check.IsNil)
-		testTar, err := os.Create(filepath.Join(tmpDir, "test.tar"))
-		if err != nil {
-			c.Fatalf("failed to create test.tar archive: %v", err)
-		}
-		defer testTar.Close()
-
-		tw := tar.NewWriter(testTar)
-
-		if err := tw.WriteHeader(&tar.Header{
-			Name: "test/foo",
-			Size: 2,
-		}); err != nil {
-			c.Fatalf("failed to write tar file header: %v", err)
-		}
-		if _, err := tw.Write([]byte("Hi")); err != nil {
-			c.Fatalf("failed to write tar file content: %v", err)
-		}
-		if err := tw.Close(); err != nil {
-			c.Fatalf("failed to close tar archive: %v", err)
-		}
-
-		xzCompressCmd := exec.Command("xz", "-k", "test.tar")
-		xzCompressCmd.Dir = tmpDir
-		out, _, err := runCommandWithOutput(xzCompressCmd)
-		if err != nil {
-			c.Fatal(err, out)
-		}
-
-		gzipCompressCmd := exec.Command("gzip", "test.tar.xz")
-		gzipCompressCmd.Dir = tmpDir
-		out, _, err = runCommandWithOutput(gzipCompressCmd)
-		if err != nil {
-			c.Fatal(err, out)
-		}
-
-		if err := ioutil.WriteFile(filepath.Join(tmpDir, "Dockerfile"), []byte(dockerfile), 0644); err != nil {
-			c.Fatalf("failed to open destination dockerfile: %v", err)
-		}
-		return fakeContextFromDir(tmpDir)
-	}()
-
-	defer ctx.Close()
-
-	if _, err := buildImageFromContext(name, ctx, true); err != nil {
-		c.Fatalf("build failed to complete for TestBuildAddTarXz: %v", err)
-	}
-
-}
-
-func (s *DockerSuite) TestBuildFromGIT(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildfromgit"
-	git, err := newFakeGit("repo", map[string]string{
-		"Dockerfile": `FROM busybox
-					ADD first /first
-					RUN [ -f /first ]
-					MAINTAINER docker`,
-		"first": "test git data",
-	}, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer git.Close()
-
-	_, err = buildImageFromPath(name, git.RepoURL, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	res, err := inspectField(name, "Author")
-	if err != nil {
-		c.Fatal(err)
-	}
-	if res != "docker" {
-		c.Fatalf("Maintainer should be docker, got %s", res)
-	}
-}
-
-func (s *DockerSuite) TestBuildFromGITWithContext(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildfromgit"
-	git, err := newFakeGit("repo", map[string]string{
-		"docker/Dockerfile": `FROM busybox
-					ADD first /first
-					RUN [ -f /first ]
-					MAINTAINER docker`,
-		"docker/first": "test git data",
-	}, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer git.Close()
-
-	u := fmt.Sprintf("%s#master:docker", git.RepoURL)
-	_, err = buildImageFromPath(name, u, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	res, err := inspectField(name, "Author")
-	if err != nil {
-		c.Fatal(err)
-	}
-	if res != "docker" {
-		c.Fatalf("Maintainer should be docker, got %s", res)
-	}
-}
-
-func (s *DockerSuite) TestBuildFromGITwithF(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildfromgitwithf"
-	git, err := newFakeGit("repo", map[string]string{
-		"myApp/myDockerfile": `FROM busybox
-					RUN echo hi from Dockerfile`,
-	}, true)
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer git.Close()
-
-	out, _, err := dockerCmdWithError("build", "-t", name, "--no-cache", "-f", "myApp/myDockerfile", git.RepoURL)
-	if err != nil {
-		c.Fatalf("Error on build. Out: %s\nErr: %v", out, err)
-	}
-
-	if !strings.Contains(out, "hi from Dockerfile") {
-		c.Fatalf("Missing expected output, got:\n%s", out)
-	}
-}
-
-func (s *DockerSuite) TestBuildFromRemoteTarball(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildfromremotetarball"
-
-	buffer := new(bytes.Buffer)
-	tw := tar.NewWriter(buffer)
-	defer tw.Close()
-
-	dockerfile := []byte(`FROM busybox
-					MAINTAINER docker`)
-	if err := tw.WriteHeader(&tar.Header{
-		Name: "Dockerfile",
-		Size: int64(len(dockerfile)),
-	}); err != nil {
-		c.Fatalf("failed to write tar file header: %v", err)
-	}
-	if _, err := tw.Write(dockerfile); err != nil {
-		c.Fatalf("failed to write tar file content: %v", err)
-	}
-	if err := tw.Close(); err != nil {
-		c.Fatalf("failed to close tar archive: %v", err)
-	}
-
-	server, err := fakeBinaryStorage(map[string]*bytes.Buffer{
-		"testT.tar": buffer,
-	})
-	c.Assert(err, check.IsNil)
-
-	defer server.Close()
-
-	_, err = buildImageFromPath(name, server.URL()+"/testT.tar", true)
-	c.Assert(err, check.IsNil)
-
-	res, err := inspectField(name, "Author")
-	c.Assert(err, check.IsNil)
-
-	if res != "docker" {
-		c.Fatalf("Maintainer should be docker, got %s", res)
-	}
-}
-
 func (s *DockerSuite) TestBuildCleanupCmdOnEntrypoint(c *check.C) {
 	testRequires(c, DaemonIsLinux)
 	name := "testbuildcmdcleanuponentrypoint"
@@ -4588,23 +1141,6 @@ func (s *DockerSuite) TestBuildEmptyCmd(c *check.C) {
 	}
 	if res != "null" {
 		c.Fatalf("Cmd %s, expected %s", res, "null")
-	}
-}
-
-func (s *DockerSuite) TestBuildOnBuildOutput(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildonbuildparent"
-	if _, err := buildImage(name, "FROM busybox\nONBUILD RUN echo foo\n", true); err != nil {
-		c.Fatal(err)
-	}
-
-	_, out, err := buildImageWithOut(name, "FROM "+name+"\nMAINTAINER quux\n", true)
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	if !strings.Contains(out, "# Executing 1 build trigger") {
-		c.Fatal("failed to find the build trigger output", out)
 	}
 }
 
@@ -4996,58 +1532,6 @@ RUN [ $(ls -l /test | awk '{print $3":"$4}') = 'root:root' ]
 
 }
 
-func (s *DockerSuite) TestBuildSymlinkBreakout(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildsymlinkbreakout"
-	tmpdir, err := ioutil.TempDir("", name)
-	c.Assert(err, check.IsNil)
-	defer os.RemoveAll(tmpdir)
-	ctx := filepath.Join(tmpdir, "context")
-	if err := os.MkdirAll(ctx, 0755); err != nil {
-		c.Fatal(err)
-	}
-	if err := ioutil.WriteFile(filepath.Join(ctx, "Dockerfile"), []byte(`
-	from busybox
-	add symlink.tar /
-	add inject /symlink/
-	`), 0644); err != nil {
-		c.Fatal(err)
-	}
-	inject := filepath.Join(ctx, "inject")
-	if err := ioutil.WriteFile(inject, nil, 0644); err != nil {
-		c.Fatal(err)
-	}
-	f, err := os.Create(filepath.Join(ctx, "symlink.tar"))
-	if err != nil {
-		c.Fatal(err)
-	}
-	w := tar.NewWriter(f)
-	w.WriteHeader(&tar.Header{
-		Name:     "symlink2",
-		Typeflag: tar.TypeSymlink,
-		Linkname: "/../../../../../../../../../../../../../../",
-		Uid:      os.Getuid(),
-		Gid:      os.Getgid(),
-	})
-	w.WriteHeader(&tar.Header{
-		Name:     "symlink",
-		Typeflag: tar.TypeSymlink,
-		Linkname: filepath.Join("symlink2", tmpdir),
-		Uid:      os.Getuid(),
-		Gid:      os.Getgid(),
-	})
-	w.Close()
-	f.Close()
-	if _, err := buildImageFromContext(name, fakeContextFromDir(ctx), false); err != nil {
-		c.Fatal(err)
-	}
-	if _, err := os.Lstat(filepath.Join(tmpdir, "inject")); err == nil {
-		c.Fatal("symlink breakout - inject")
-	} else if !os.IsNotExist(err) {
-		c.Fatalf("unexpected error: %v", err)
-	}
-}
-
 func (s *DockerSuite) TestBuildXZHost(c *check.C) {
 	// /usr/local/sbin/xz gets permission denied for the user
 	testRequires(c, NotUserNamespace)
@@ -5349,64 +1833,6 @@ func (s *DockerSuite) TestBuildFromOfficialNames(c *check.C) {
 			c.Errorf("Build failed using FROM %s: %s", fromName, err)
 		}
 		deleteImages(imgName)
-	}
-}
-
-func (s *DockerSuite) TestBuildDockerfileOutsideContext(c *check.C) {
-	testRequires(c, UnixCli) // uses os.Symlink: not implemented in windows at the time of writing (go-1.4.2)
-	testRequires(c, DaemonIsLinux)
-
-	name := "testbuilddockerfileoutsidecontext"
-	tmpdir, err := ioutil.TempDir("", name)
-	c.Assert(err, check.IsNil)
-	defer os.RemoveAll(tmpdir)
-	ctx := filepath.Join(tmpdir, "context")
-	if err := os.MkdirAll(ctx, 0755); err != nil {
-		c.Fatal(err)
-	}
-	if err := ioutil.WriteFile(filepath.Join(ctx, "Dockerfile"), []byte("FROM scratch\nENV X Y"), 0644); err != nil {
-		c.Fatal(err)
-	}
-	wd, err := os.Getwd()
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer os.Chdir(wd)
-	if err := os.Chdir(ctx); err != nil {
-		c.Fatal(err)
-	}
-	if err := ioutil.WriteFile(filepath.Join(tmpdir, "outsideDockerfile"), []byte("FROM scratch\nENV x y"), 0644); err != nil {
-		c.Fatal(err)
-	}
-	if err := os.Symlink(filepath.Join("..", "outsideDockerfile"), filepath.Join(ctx, "dockerfile1")); err != nil {
-		c.Fatal(err)
-	}
-	if err := os.Symlink(filepath.Join(tmpdir, "outsideDockerfile"), filepath.Join(ctx, "dockerfile2")); err != nil {
-		c.Fatal(err)
-	}
-
-	for _, dockerfilePath := range []string{
-		filepath.Join("..", "outsideDockerfile"),
-		filepath.Join(ctx, "dockerfile1"),
-		filepath.Join(ctx, "dockerfile2"),
-	} {
-		out, _, err := dockerCmdWithError("build", "-t", name, "--no-cache", "-f", dockerfilePath, ".")
-		if err == nil {
-			c.Fatalf("Expected error with %s. Out: %s", dockerfilePath, out)
-		}
-		if !strings.Contains(out, "must be within the build context") && !strings.Contains(out, "Cannot locate Dockerfile") {
-			c.Fatalf("Unexpected error with %s. Out: %s", dockerfilePath, out)
-		}
-		deleteImages(name)
-	}
-
-	os.Chdir(tmpdir)
-
-	// Path to Dockerfile should be resolved relative to working directory, not relative to context.
-	// There is a Dockerfile in the context, but since there is no Dockerfile in the current directory, the following should fail
-	out, _, err := dockerCmdWithError("build", "-t", name, "--no-cache", "-f", "Dockerfile", ctx)
-	if err == nil {
-		c.Fatalf("Expected error. Out: %s", out)
 	}
 }
 
@@ -5810,43 +2236,6 @@ func (s *DockerTrustSuite) TestTrustedBuildUntrustedTag(c *check.C) {
 	}
 }
 
-func (s *DockerTrustSuite) TestBuildContextDirIsSymlink(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	tempDir, err := ioutil.TempDir("", "test-build-dir-is-symlink-")
-	c.Assert(err, check.IsNil)
-	defer os.RemoveAll(tempDir)
-
-	// Make a real context directory in this temp directory with a simple
-	// Dockerfile.
-	realContextDirname := filepath.Join(tempDir, "context")
-	if err := os.Mkdir(realContextDirname, os.FileMode(0755)); err != nil {
-		c.Fatal(err)
-	}
-
-	if err = ioutil.WriteFile(
-		filepath.Join(realContextDirname, "Dockerfile"),
-		[]byte(`
-			FROM busybox
-			RUN echo hello world
-		`),
-		os.FileMode(0644),
-	); err != nil {
-		c.Fatal(err)
-	}
-
-	// Make a symlink to the real context directory.
-	contextSymlinkName := filepath.Join(tempDir, "context_link")
-	if err := os.Symlink(realContextDirname, contextSymlinkName); err != nil {
-		c.Fatal(err)
-	}
-
-	// Executing the build with the symlink as the specified context should
-	// *not* fail.
-	if out, exitStatus := dockerCmd(c, "build", contextSymlinkName); exitStatus != 0 {
-		c.Fatalf("build failed with exit status %d: %s", exitStatus, out)
-	}
-}
-
 // Issue #15634: COPY fails when path starts with "null"
 func (s *DockerSuite) TestBuildNullStringInAddCopyVolume(c *check.C) {
 	testRequires(c, DaemonIsLinux)
@@ -5884,527 +2273,6 @@ func (s *DockerSuite) TestBuildStopSignal(c *check.C) {
 
 	if res != `"SIGKILL"` {
 		c.Fatalf("Signal %s, expected SIGKILL", res)
-	}
-}
-
-func (s *DockerSuite) TestBuildBuildTimeArg(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	imgName := "bldargtest"
-	envKey := "foo"
-	envVal := "bar"
-	args := []string{
-		"--build-arg", fmt.Sprintf("%s=%s", envKey, envVal),
-	}
-	dockerfile := fmt.Sprintf(`FROM busybox
-		ARG %s
-		RUN echo $%s
-		CMD echo $%s`, envKey, envKey, envKey)
-
-	if _, out, err := buildImageWithOut(imgName, dockerfile, true, args...); err != nil || !strings.Contains(out, envVal) {
-		if err != nil {
-			c.Fatalf("build failed to complete: %q %q", out, err)
-		}
-		c.Fatalf("failed to access environment variable in output: %q expected: %q", out, envVal)
-	}
-
-	containerName := "bldargCont"
-	if out, _ := dockerCmd(c, "run", "--name", containerName, imgName); out != "\n" {
-		c.Fatalf("run produced invalid output: %q, expected empty string", out)
-	}
-}
-
-func (s *DockerSuite) TestBuildBuildTimeArgHistory(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	imgName := "bldargtest"
-	envKey := "foo"
-	envVal := "bar"
-	envDef := "bar1"
-	args := []string{
-		"--build-arg", fmt.Sprintf("%s=%s", envKey, envVal),
-	}
-	dockerfile := fmt.Sprintf(`FROM busybox
-		ARG %s=%s`, envKey, envDef)
-
-	if _, out, err := buildImageWithOut(imgName, dockerfile, true, args...); err != nil || !strings.Contains(out, envVal) {
-		if err != nil {
-			c.Fatalf("build failed to complete: %q %q", out, err)
-		}
-		c.Fatalf("failed to access environment variable in output: %q expected: %q", out, envVal)
-	}
-
-	out, _ := dockerCmd(c, "history", "--no-trunc", imgName)
-	outputTabs := strings.Split(out, "\n")[1]
-	if !strings.Contains(outputTabs, envDef) {
-		c.Fatalf("failed to find arg default in image history output: %q expected: %q", outputTabs, envDef)
-	}
-}
-
-func (s *DockerSuite) TestBuildBuildTimeArgCacheHit(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	imgName := "bldargtest"
-	envKey := "foo"
-	envVal := "bar"
-	args := []string{
-		"--build-arg", fmt.Sprintf("%s=%s", envKey, envVal),
-	}
-	dockerfile := fmt.Sprintf(`FROM busybox
-		ARG %s
-		RUN echo $%s`, envKey, envKey)
-
-	origImgID := ""
-	var err error
-	if origImgID, err = buildImage(imgName, dockerfile, true, args...); err != nil {
-		c.Fatal(err)
-	}
-
-	imgNameCache := "bldargtestcachehit"
-	if newImgID, err := buildImage(imgNameCache, dockerfile, true, args...); err != nil || newImgID != origImgID {
-		if err != nil {
-			c.Fatal(err)
-		}
-		c.Fatalf("build didn't use cache! expected image id: %q built image id: %q", origImgID, newImgID)
-	}
-}
-
-func (s *DockerSuite) TestBuildBuildTimeArgCacheMissExtraArg(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	imgName := "bldargtest"
-	envKey := "foo"
-	envVal := "bar"
-	extraEnvKey := "foo1"
-	extraEnvVal := "bar1"
-	args := []string{
-		"--build-arg", fmt.Sprintf("%s=%s", envKey, envVal),
-	}
-
-	dockerfile := fmt.Sprintf(`FROM busybox
-		ARG %s
-		ARG %s
-		RUN echo $%s`, envKey, extraEnvKey, envKey)
-
-	origImgID := ""
-	var err error
-	if origImgID, err = buildImage(imgName, dockerfile, true, args...); err != nil {
-		c.Fatal(err)
-	}
-
-	imgNameCache := "bldargtestcachemiss"
-	args = append(args, "--build-arg", fmt.Sprintf("%s=%s", extraEnvKey, extraEnvVal))
-	if newImgID, err := buildImage(imgNameCache, dockerfile, true, args...); err != nil || newImgID == origImgID {
-		if err != nil {
-			c.Fatal(err)
-		}
-		c.Fatalf("build used cache, expected a miss!")
-	}
-}
-
-func (s *DockerSuite) TestBuildBuildTimeArgCacheMissSameArgDiffVal(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	imgName := "bldargtest"
-	envKey := "foo"
-	envVal := "bar"
-	newEnvVal := "bar1"
-	args := []string{
-		"--build-arg", fmt.Sprintf("%s=%s", envKey, envVal),
-	}
-
-	dockerfile := fmt.Sprintf(`FROM busybox
-		ARG %s
-		RUN echo $%s`, envKey, envKey)
-
-	origImgID := ""
-	var err error
-	if origImgID, err = buildImage(imgName, dockerfile, true, args...); err != nil {
-		c.Fatal(err)
-	}
-
-	imgNameCache := "bldargtestcachemiss"
-	args = []string{
-		"--build-arg", fmt.Sprintf("%s=%s", envKey, newEnvVal),
-	}
-	if newImgID, err := buildImage(imgNameCache, dockerfile, true, args...); err != nil || newImgID == origImgID {
-		if err != nil {
-			c.Fatal(err)
-		}
-		c.Fatalf("build used cache, expected a miss!")
-	}
-}
-
-func (s *DockerSuite) TestBuildBuildTimeArgOverrideArgDefinedBeforeEnv(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	imgName := "bldargtest"
-	envKey := "foo"
-	envVal := "bar"
-	envValOveride := "barOverride"
-	args := []string{
-		"--build-arg", fmt.Sprintf("%s=%s", envKey, envVal),
-	}
-	dockerfile := fmt.Sprintf(`FROM busybox
-		ARG %s
-		ENV %s %s
-		RUN echo $%s
-		CMD echo $%s
-        `, envKey, envKey, envValOveride, envKey, envKey)
-
-	if _, out, err := buildImageWithOut(imgName, dockerfile, true, args...); err != nil || strings.Count(out, envValOveride) != 2 {
-		if err != nil {
-			c.Fatalf("build failed to complete: %q %q", out, err)
-		}
-		c.Fatalf("failed to access environment variable in output: %q expected: %q", out, envValOveride)
-	}
-
-	containerName := "bldargCont"
-	if out, _ := dockerCmd(c, "run", "--name", containerName, imgName); !strings.Contains(out, envValOveride) {
-		c.Fatalf("run produced invalid output: %q, expected %q", out, envValOveride)
-	}
-}
-
-func (s *DockerSuite) TestBuildBuildTimeArgOverrideEnvDefinedBeforeArg(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	imgName := "bldargtest"
-	envKey := "foo"
-	envVal := "bar"
-	envValOveride := "barOverride"
-	args := []string{
-		"--build-arg", fmt.Sprintf("%s=%s", envKey, envVal),
-	}
-	dockerfile := fmt.Sprintf(`FROM busybox
-		ENV %s %s
-		ARG %s
-		RUN echo $%s
-		CMD echo $%s
-        `, envKey, envValOveride, envKey, envKey, envKey)
-
-	if _, out, err := buildImageWithOut(imgName, dockerfile, true, args...); err != nil || strings.Count(out, envValOveride) != 2 {
-		if err != nil {
-			c.Fatalf("build failed to complete: %q %q", out, err)
-		}
-		c.Fatalf("failed to access environment variable in output: %q expected: %q", out, envValOveride)
-	}
-
-	containerName := "bldargCont"
-	if out, _ := dockerCmd(c, "run", "--name", containerName, imgName); !strings.Contains(out, envValOveride) {
-		c.Fatalf("run produced invalid output: %q, expected %q", out, envValOveride)
-	}
-}
-
-func (s *DockerSuite) TestBuildBuildTimeArgExpansion(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	imgName := "bldvarstest"
-
-	wdVar := "WDIR"
-	wdVal := "/tmp/"
-	addVar := "AFILE"
-	addVal := "addFile"
-	copyVar := "CFILE"
-	copyVal := "copyFile"
-	envVar := "foo"
-	envVal := "bar"
-	exposeVar := "EPORT"
-	exposeVal := "9999"
-	userVar := "USER"
-	userVal := "testUser"
-	volVar := "VOL"
-	volVal := "/testVol/"
-	args := []string{
-		"--build-arg", fmt.Sprintf("%s=%s", wdVar, wdVal),
-		"--build-arg", fmt.Sprintf("%s=%s", addVar, addVal),
-		"--build-arg", fmt.Sprintf("%s=%s", copyVar, copyVal),
-		"--build-arg", fmt.Sprintf("%s=%s", envVar, envVal),
-		"--build-arg", fmt.Sprintf("%s=%s", exposeVar, exposeVal),
-		"--build-arg", fmt.Sprintf("%s=%s", userVar, userVal),
-		"--build-arg", fmt.Sprintf("%s=%s", volVar, volVal),
-	}
-	ctx, err := fakeContext(fmt.Sprintf(`FROM busybox
-		ARG %s
-		WORKDIR ${%s}
-		ARG %s
-		ADD ${%s} testDir/
-		ARG %s
-		COPY $%s testDir/
-		ARG %s
-		ENV %s=${%s}
-		ARG %s
-		EXPOSE $%s
-		ARG %s
-		USER $%s
-		ARG %s
-		VOLUME ${%s}`,
-		wdVar, wdVar, addVar, addVar, copyVar, copyVar, envVar, envVar,
-		envVar, exposeVar, exposeVar, userVar, userVar, volVar, volVar),
-		map[string]string{
-			addVal:  "some stuff",
-			copyVal: "some stuff",
-		})
-	if err != nil {
-		c.Fatal(err)
-	}
-	defer ctx.Close()
-
-	if _, err := buildImageFromContext(imgName, ctx, true, args...); err != nil {
-		c.Fatal(err)
-	}
-
-	var resMap map[string]interface{}
-	var resArr []string
-	res := ""
-	res, err = inspectField(imgName, "Config.WorkingDir")
-	if err != nil {
-		c.Fatal(err)
-	}
-	if res != wdVal {
-		c.Fatalf("Config.WorkingDir value mismatch. Expected: %s, got: %s", wdVal, res)
-	}
-
-	err = inspectFieldAndMarshall(imgName, "Config.Env", &resArr)
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	found := false
-	for _, v := range resArr {
-		if fmt.Sprintf("%s=%s", envVar, envVal) == v {
-			found = true
-			break
-		}
-	}
-	if !found {
-		c.Fatalf("Config.Env value mismatch. Expected <key=value> to exist: %s=%s, got: %v",
-			envVar, envVal, resArr)
-	}
-
-	err = inspectFieldAndMarshall(imgName, "Config.ExposedPorts", &resMap)
-	if err != nil {
-		c.Fatal(err)
-	}
-	if _, ok := resMap[fmt.Sprintf("%s/tcp", exposeVal)]; !ok {
-		c.Fatalf("Config.ExposedPorts value mismatch. Expected exposed port: %s/tcp, got: %v", exposeVal, resMap)
-	}
-
-	res, err = inspectField(imgName, "Config.User")
-	if err != nil {
-		c.Fatal(err)
-	}
-	if res != userVal {
-		c.Fatalf("Config.User value mismatch. Expected: %s, got: %s", userVal, res)
-	}
-
-	err = inspectFieldAndMarshall(imgName, "Config.Volumes", &resMap)
-	if err != nil {
-		c.Fatal(err)
-	}
-	if _, ok := resMap[volVal]; !ok {
-		c.Fatalf("Config.Volumes value mismatch. Expected volume: %s, got: %v", volVal, resMap)
-	}
-}
-
-func (s *DockerSuite) TestBuildBuildTimeArgExpansionOverride(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	imgName := "bldvarstest"
-	envKey := "foo"
-	envVal := "bar"
-	envKey1 := "foo1"
-	envValOveride := "barOverride"
-	args := []string{
-		"--build-arg", fmt.Sprintf("%s=%s", envKey, envVal),
-	}
-	dockerfile := fmt.Sprintf(`FROM busybox
-		ARG %s
-		ENV %s %s
-		ENV %s ${%s}
-		RUN echo $%s
-		CMD echo $%s`, envKey, envKey, envValOveride, envKey1, envKey, envKey1, envKey1)
-
-	if _, out, err := buildImageWithOut(imgName, dockerfile, true, args...); err != nil || strings.Count(out, envValOveride) != 2 {
-		if err != nil {
-			c.Fatalf("build failed to complete: %q %q", out, err)
-		}
-		c.Fatalf("failed to access environment variable in output: %q expected: %q", out, envValOveride)
-	}
-
-	containerName := "bldargCont"
-	if out, _ := dockerCmd(c, "run", "--name", containerName, imgName); !strings.Contains(out, envValOveride) {
-		c.Fatalf("run produced invalid output: %q, expected %q", out, envValOveride)
-	}
-}
-
-func (s *DockerSuite) TestBuildBuildTimeArgUntrustedDefinedAfterUse(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	imgName := "bldargtest"
-	envKey := "foo"
-	envVal := "bar"
-	args := []string{
-		"--build-arg", fmt.Sprintf("%s=%s", envKey, envVal),
-	}
-	dockerfile := fmt.Sprintf(`FROM busybox
-		RUN echo $%s
-		ARG %s
-		CMD echo $%s`, envKey, envKey, envKey)
-
-	if _, out, err := buildImageWithOut(imgName, dockerfile, true, args...); err != nil || strings.Contains(out, envVal) {
-		if err != nil {
-			c.Fatalf("build failed to complete: %q %q", out, err)
-		}
-		c.Fatalf("able to access environment variable in output: %q expected to be missing", out)
-	}
-
-	containerName := "bldargCont"
-	if out, _ := dockerCmd(c, "run", "--name", containerName, imgName); out != "\n" {
-		c.Fatalf("run produced invalid output: %q, expected empty string", out)
-	}
-}
-
-func (s *DockerSuite) TestBuildBuildTimeArgBuiltinArg(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	imgName := "bldargtest"
-	envKey := "HTTP_PROXY"
-	envVal := "bar"
-	args := []string{
-		"--build-arg", fmt.Sprintf("%s=%s", envKey, envVal),
-	}
-	dockerfile := fmt.Sprintf(`FROM busybox
-		RUN echo $%s
-		CMD echo $%s`, envKey, envKey)
-
-	if _, out, err := buildImageWithOut(imgName, dockerfile, true, args...); err != nil || !strings.Contains(out, envVal) {
-		if err != nil {
-			c.Fatalf("build failed to complete: %q %q", out, err)
-		}
-		c.Fatalf("failed to access environment variable in output: %q expected: %q", out, envVal)
-	}
-
-	containerName := "bldargCont"
-	if out, _ := dockerCmd(c, "run", "--name", containerName, imgName); out != "\n" {
-		c.Fatalf("run produced invalid output: %q, expected empty string", out)
-	}
-}
-
-func (s *DockerSuite) TestBuildBuildTimeArgDefaultOverride(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	imgName := "bldargtest"
-	envKey := "foo"
-	envVal := "bar"
-	envValOveride := "barOverride"
-	args := []string{
-		"--build-arg", fmt.Sprintf("%s=%s", envKey, envValOveride),
-	}
-	dockerfile := fmt.Sprintf(`FROM busybox
-		ARG %s=%s
-		ENV %s $%s
-		RUN echo $%s
-		CMD echo $%s`, envKey, envVal, envKey, envKey, envKey, envKey)
-
-	if _, out, err := buildImageWithOut(imgName, dockerfile, true, args...); err != nil || strings.Count(out, envValOveride) != 1 {
-		if err != nil {
-			c.Fatalf("build failed to complete: %q %q", out, err)
-		}
-		c.Fatalf("failed to access environment variable in output: %q expected: %q", out, envValOveride)
-	}
-
-	containerName := "bldargCont"
-	if out, _ := dockerCmd(c, "run", "--name", containerName, imgName); !strings.Contains(out, envValOveride) {
-		c.Fatalf("run produced invalid output: %q, expected %q", out, envValOveride)
-	}
-}
-
-func (s *DockerSuite) TestBuildBuildTimeArgMultiArgsSameLine(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	imgName := "bldargtest"
-	envKey := "foo"
-	envKey1 := "foo1"
-	args := []string{}
-	dockerfile := fmt.Sprintf(`FROM busybox
-		ARG %s %s`, envKey, envKey1)
-
-	errStr := "ARG requires exactly one argument definition"
-	if _, out, err := buildImageWithOut(imgName, dockerfile, true, args...); err == nil {
-		c.Fatalf("build succeeded, expected to fail. Output: %v", out)
-	} else if !strings.Contains(out, errStr) {
-		c.Fatalf("Unexpected error. output: %q, expected error: %q", out, errStr)
-	}
-}
-
-func (s *DockerSuite) TestBuildBuildTimeArgUnconsumedArg(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	imgName := "bldargtest"
-	envKey := "foo"
-	envVal := "bar"
-	args := []string{
-		"--build-arg", fmt.Sprintf("%s=%s", envKey, envVal),
-	}
-	dockerfile := fmt.Sprintf(`FROM busybox
-		RUN echo $%s
-		CMD echo $%s`, envKey, envKey)
-
-	errStr := "One or more build-args"
-	if _, out, err := buildImageWithOut(imgName, dockerfile, true, args...); err == nil {
-		c.Fatalf("build succeeded, expected to fail. Output: %v", out)
-	} else if !strings.Contains(out, errStr) {
-		c.Fatalf("Unexpected error. output: %q, expected error: %q", out, errStr)
-	}
-
-}
-
-func (s *DockerSuite) TestBuildBuildTimeArgQuotedValVariants(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	imgName := "bldargtest"
-	envKey := "foo"
-	envKey1 := "foo1"
-	envKey2 := "foo2"
-	envKey3 := "foo3"
-	args := []string{}
-	dockerfile := fmt.Sprintf(`FROM busybox
-		ARG %s=""
-		ARG %s=''
-		ARG %s="''"
-		ARG %s='""'
-		RUN [ "$%s" != "$%s" ]
-		RUN [ "$%s" != "$%s" ]
-		RUN [ "$%s" != "$%s" ]
-		RUN [ "$%s" != "$%s" ]
-		RUN [ "$%s" != "$%s" ]`, envKey, envKey1, envKey2, envKey3,
-		envKey, envKey2, envKey, envKey3, envKey1, envKey2, envKey1, envKey3,
-		envKey2, envKey3)
-
-	if _, out, err := buildImageWithOut(imgName, dockerfile, true, args...); err != nil {
-		c.Fatalf("build failed to complete: %q %q", out, err)
-	}
-}
-
-func (s *DockerSuite) TestBuildBuildTimeArgEmptyValVariants(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	imgName := "bldargtest"
-	envKey := "foo"
-	envKey1 := "foo1"
-	envKey2 := "foo2"
-	args := []string{}
-	dockerfile := fmt.Sprintf(`FROM busybox
-		ARG %s=
-		ARG %s=""
-		ARG %s=''
-		RUN [ "$%s" == "$%s" ]
-		RUN [ "$%s" == "$%s" ]
-		RUN [ "$%s" == "$%s" ]`, envKey, envKey1, envKey2, envKey, envKey1, envKey1, envKey2, envKey, envKey2)
-
-	if _, out, err := buildImageWithOut(imgName, dockerfile, true, args...); err != nil {
-		c.Fatalf("build failed to complete: %q %q", out, err)
-	}
-}
-
-func (s *DockerSuite) TestBuildBuildTimeArgDefintionWithNoEnvInjection(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	imgName := "bldargtest"
-	envKey := "foo"
-	args := []string{}
-	dockerfile := fmt.Sprintf(`FROM busybox
-		ARG %s
-		RUN env`, envKey)
-
-	if _, out, err := buildImageWithOut(imgName, dockerfile, true, args...); err != nil || strings.Count(out, envKey) != 1 {
-		if err != nil {
-			c.Fatalf("build failed to complete: %q %q", out, err)
-		}
-		c.Fatalf("unexpected number of occurrences of the arg in output: %q expected: 1", out)
 	}
 }
 
@@ -6471,130 +2339,6 @@ func (s *DockerSuite) TestBuildMultipleTags(c *check.C) {
 	id2, err := getIDByName("tag2:v2")
 	c.Assert(err, check.IsNil)
 	c.Assert(id1, check.Equals, id2)
-}
-
-// #17290
-func (s *DockerSuite) TestBuildCacheBrokenSymlink(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildbrokensymlink"
-	ctx, err := fakeContext(`
-	FROM busybox
-	COPY . ./`,
-		map[string]string{
-			"foo": "bar",
-		})
-	c.Assert(err, checker.IsNil)
-	defer ctx.Close()
-
-	err = os.Symlink(filepath.Join(ctx.Dir, "nosuchfile"), filepath.Join(ctx.Dir, "asymlink"))
-	c.Assert(err, checker.IsNil)
-
-	// warm up cache
-	_, err = buildImageFromContext(name, ctx, true)
-	c.Assert(err, checker.IsNil)
-
-	// add new file to context, should invalidate cache
-	err = ioutil.WriteFile(filepath.Join(ctx.Dir, "newfile"), []byte("foo"), 0644)
-	c.Assert(err, checker.IsNil)
-
-	_, out, err := buildImageFromContextWithOut(name, ctx, true)
-	c.Assert(err, checker.IsNil)
-
-	c.Assert(out, checker.Not(checker.Contains), "Using cache")
-
-}
-
-func (s *DockerSuite) TestBuildFollowSymlinkToFile(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildbrokensymlink"
-	ctx, err := fakeContext(`
-	FROM busybox
-	COPY asymlink target`,
-		map[string]string{
-			"foo": "bar",
-		})
-	c.Assert(err, checker.IsNil)
-	defer ctx.Close()
-
-	err = os.Symlink("foo", filepath.Join(ctx.Dir, "asymlink"))
-	c.Assert(err, checker.IsNil)
-
-	id, err := buildImageFromContext(name, ctx, true)
-	c.Assert(err, checker.IsNil)
-
-	out, _ := dockerCmd(c, "run", "--rm", id, "cat", "target")
-	c.Assert(out, checker.Matches, "bar")
-
-	// change target file should invalidate cache
-	err = ioutil.WriteFile(filepath.Join(ctx.Dir, "foo"), []byte("baz"), 0644)
-	c.Assert(err, checker.IsNil)
-
-	id, out, err = buildImageFromContextWithOut(name, ctx, true)
-	c.Assert(err, checker.IsNil)
-	c.Assert(out, checker.Not(checker.Contains), "Using cache")
-
-	out, _ = dockerCmd(c, "run", "--rm", id, "cat", "target")
-	c.Assert(out, checker.Matches, "baz")
-}
-
-func (s *DockerSuite) TestBuildFollowSymlinkToDir(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildbrokensymlink"
-	ctx, err := fakeContext(`
-	FROM busybox
-	COPY asymlink /`,
-		map[string]string{
-			"foo/abc": "bar",
-			"foo/def": "baz",
-		})
-	c.Assert(err, checker.IsNil)
-	defer ctx.Close()
-
-	err = os.Symlink("foo", filepath.Join(ctx.Dir, "asymlink"))
-	c.Assert(err, checker.IsNil)
-
-	id, err := buildImageFromContext(name, ctx, true)
-	c.Assert(err, checker.IsNil)
-
-	out, _ := dockerCmd(c, "run", "--rm", id, "cat", "abc", "def")
-	c.Assert(out, checker.Matches, "barbaz")
-
-	// change target file should invalidate cache
-	err = ioutil.WriteFile(filepath.Join(ctx.Dir, "foo/def"), []byte("bax"), 0644)
-	c.Assert(err, checker.IsNil)
-
-	id, out, err = buildImageFromContextWithOut(name, ctx, true)
-	c.Assert(err, checker.IsNil)
-	c.Assert(out, checker.Not(checker.Contains), "Using cache")
-
-	out, _ = dockerCmd(c, "run", "--rm", id, "cat", "abc", "def")
-	c.Assert(out, checker.Matches, "barbax")
-
-}
-
-// TestBuildSymlinkBasename tests that target file gets basename from symlink,
-// not from the target file.
-func (s *DockerSuite) TestBuildSymlinkBasename(c *check.C) {
-	testRequires(c, DaemonIsLinux)
-	name := "testbuildbrokensymlink"
-	ctx, err := fakeContext(`
-	FROM busybox
-	COPY asymlink /`,
-		map[string]string{
-			"foo": "bar",
-		})
-	c.Assert(err, checker.IsNil)
-	defer ctx.Close()
-
-	err = os.Symlink("foo", filepath.Join(ctx.Dir, "asymlink"))
-	c.Assert(err, checker.IsNil)
-
-	id, err := buildImageFromContext(name, ctx, true)
-	c.Assert(err, checker.IsNil)
-
-	out, _ := dockerCmd(c, "run", "--rm", id, "cat", "asymlink")
-	c.Assert(out, checker.Matches, "bar")
-
 }
 
 // #17827
